Image and video samples

Rvision provides a few image and video samples for illustrating the examples of its different functions. Below are references to the original material from which these samples have been extracted.

References

"Big Buck Bunny", © copyright 2008, Blender Foundation / www.bigbuckbunny.org - Distributed under the Creative Commons Attribution 3.0 license.

EC Funded CAVIAR project/IST 2001 37540 - http://homepages.inf.ed.ac.uk/rbf/CAVIAR/

Pixabay / https://pixabay.com/en/dots-blue-background-fabric-313325/ - Distributed under the Creative Commons CC0 Public Domain license.