An S4 Class Containing an OpenCV Video

A Video object contains an OpenCV video that originates from a video file.

Function for creating Video objects from video files.

Usage

video(filename, api = "ANY")

Arguments

filename

An character string corresponding to the path to a video file.

api

A character string corresponding to the API to use for reading the video from the file (see Note; default: "ANY").

Value

A Video object.

Slots

dim,ncol,nrow

Functions returning the dimensions of the object.

nframes

Function returning the number of frames in the object.

codec

Function returning the codec of the object.

fps

Function returning the frame rate of the object.

frame

Function returning the index of the frame to be read next.

get,set

Functions to access and set internal properties of the object.

isOpened

Function to check the status of the video stream.

readNext,readFrame

Functions to access the next or an arbitrary frame in the stream.

release

Function to release the object from memory.

Note

Hereafter is a list of all supported APIs. Note that not all APIs will be available on your computer (actually most of them will not be).

"ANY":

automatically select an API.

"VFW":

Video For Windows.

"V4L", "V4L2":

Video For Linux.

"FIREWIRE", "FIREWARE", "IEEE1394", "DC1394", "CMU1394":

IEEE 1394 drivers.

"QT":

Quicktime.

"UNICAP":

Unicap drivers.

"DSHOW":

DirectShow.

"PVAPI":

PvAPI, Prosilica GigE SDK.

"OPENNI":

OpenNI (for Kinect).

"OPENNI_ASUS":

OpenNI (for Asus Xtion).

"XIAPI":

XIMEA Camera API.

"AVFOUNDATION":

AVFoundation framework for iOS and OSX > Lion.

"GIGANETIX":

Smartek Giganetix GigEVisionSDK.

"MSMF":

Microsoft Media Foundation.

"WINRT":

Microsoft Windows Runtime using Media Foundation.

"INTELPERC":

Intel Perceptual Computing SDK.

"OPENNI2":

OpenNI2 (for Kinect).

"OPENNI2_ASUS":

OpenNI2 (for Asus Xtion and Occipital Structure sensors).

"GPHOTO2":

gPhoto2 connection.

"GSTREAMER":

GStreamer.

"FFMPEG":

FFMPEG library.

"IMAGES":

OpenCV Image Sequence.

"ARAVIS":

Aravis SDK.

"OPENCV_MJPEG":

Built-in OpenCV MotionJPEG codec.

"INTEL_MFX":

Intel MediaSDK.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))