Find the Convexity Defects of a Contour

convexityDefects finds the convexity defects of a contour, that is area that do not belong to an object but are located inside of its convex hull.

convexityDefects(contours, id = NULL)

Arguments

contours
A list of two matrices as produced by findContours.

id
An optional vector indicating the identity of the specific contours for which to run the function.

Value

A matrix with 5 columns:

"id": the contour identity (indicates the set of points belonging to the same contour).

"start_index": index of the first point of the contour belonging to a convexity defect.

"end_index": index of the last point of the contour belonging to a convexity defect.

"farthest_pt_index": index of the point of the contour belonging to a convexity defect and that is the farthest away from the convex hull.

"fixpt_depth": distance between the farthest contour point and the convex hull.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

dots <- image(system.file("sample_img/dots.jpg", package = "Rvision"))
dots_gray <- changeColorSpace(dots, "GRAY")
dots_bin <- dots_gray < 200
contours <- findContours(dots_bin)
convexityDefects(contours, id = c(3, 5))

#>       id start_index end_index farthest_pt_index fixpt_depth
#>  [1,]  3           1         2                 2         217
#>  [2,]  3           3         4                 4         162
#>  [3,]  3           9        10                10         114
#>  [4,]  3          11        12                12         176
#>  [5,]  3          15        16                16         154
#>  [6,]  3          18        19                19         154
#>  [7,]  3          20        21                21         142
#>  [8,]  3          22        23                23         114
#>  [9,]  3          28        29                29         162
#> [10,]  3          34        35                35         210
#> [11,]  3          37        38                38         210
#> [12,]  3          39        40                40         186
#> [13,]  3          41        42                42         162
#> [14,]  3          45        46                46         114
#> [15,]  3          53        54                54         142
#> [16,]  3          58        59                59         142
#> [17,]  3          60        63                63         176
#> [18,]  3          64        65                65         114
#> [19,]  3          66        67                67         162
#> [20,]  3          72        73                73         201
#> [21,]  3          75        76                76         201
#> [22,]  3          77        78                78         162
#> [23,]  3          81        82                82         114
#> [24,]  3          85        86                86         176
#> [25,]  3          90        91                91         142
#> [26,]  3          92        95                95         176
#> [27,]  3          96        97                97         190
#> [28,]  3         100       101               101         162
#> [29,]  3         102       103               103         186
#> [30,]  3         104       105               105         201
#> [31,]  3         107       108               108         201
#> [32,]  3         109       110               110         186
#> [33,]  3         111       112               112         162
#> [34,]  3         113       116               116         190
#> [35,]  3         117       118               118         114
#> [36,]  3         121       122               122         142
#> [37,]  3         126       127               127         142
#> [38,]  3         130       131               131         114
#> [39,]  3         132       133               133         190
#> [40,]  3         136       137               137         162
#> [41,]  3         140       141               141         222
#> [42,]  5           1         2                 2         186
#> [43,]  5           3         4                 4         162
#> [44,]  5           5         6                 6         114
#> [45,]  5           9        10                10         142
#> [46,]  5          11        12                12         154
#> [47,]  5          14        15                15         154
#> [48,]  5          16        17                17         142
#> [49,]  5          18        19                19         114
#> [50,]  5          28        29                29         162
#> [51,]  5          32        33                33         186
#> [52,]  5          36        37                37         229
#> [53,]  5          39        40                40         226
#> [54,]  5          41        42                42         186
#> [55,]  5          43        44                44         162
#> [56,]  5          47        48                48         114
#> [57,]  5          56        57                57         114
#> [58,]  5          60        61                61         162
#> [59,]  5          63        64                64         162
#> [60,]  5          65        66                66         114
#> [61,]  5          69        70                70         142
#> [62,]  5          71        72                72         154
#> [63,]  5          74        75                75         160
#> [64,]  5          76        79                79         176
#> [65,]  5          80        81                81         114
#> [66,]  5          88        89                89         162
#> [67,]  5          96        97                97         201
#> [68,]  5          98        99                99         217
#> [69,]  5         101       102               102         217
#> [70,]  5         103       104               104         186
#> [71,]  5         105       106               106         162
#> [72,]  5         107       108               108         114
#> [73,]  5         113       114               114         142
#> [74,]  5         116       117               117         142
#> [75,]  5         118       119               119         114
#> [76,]  5         122       123               123         186