Find Non-Zero Pixels in an Image

findNonZero retrieves the locations of all non-zero pixels in an image.

Usage

findNonZero(image, values = FALSE)

Arguments

image

A single-channel Image object.

values

A logical indicating whether the values of the non-zero pixels should also be returned (default: FALSE).

Value

If values=FALSE, a matrix with two columns, corresponding to the x (columns) and y (rows) coordinates of the non-zero pixels. If

values=TRUE, an additional column corresponding to the pixel values will also be returned.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_gray <- changeColorSpace(balloon, "GRAY")
findNonZero(balloon_gray)
#>             x   y
#>      [1,]   1 360
#>      [2,]   2 360
#>      [3,]   3 360
#>      [4,]   4 360
#>      [5,]   5 360
#>      [6,]   6 360
#>      [7,]   7 360
#>      [8,]   8 360
#>      [9,]   9 360
#>     [10,]  10 360
#>     [11,]  11 360
#>     [12,]  12 360
#>     [13,]  13 360
#>     [14,]  14 360
#>     [15,]  15 360
#>     [16,]  16 360
#>     [17,]  17 360
#>     [18,]  18 360
#>     [19,]  19 360
#>     [20,]  20 360
#>     [21,]  21 360
#>     [22,]  22 360
#>     [23,]  23 360
#>     [24,]  24 360
#>     [25,]  25 360
#>     [26,]  26 360
#>     [27,]  27 360
#>     [28,]  28 360
#>     [29,]  29 360
#>     [30,]  30 360
#>     [31,]  31 360
#>     [32,]  32 360
#>     [33,]  33 360
#>     [34,]  34 360
#>     [35,]  35 360
#>     [36,]  36 360
#>     [37,]  37 360
#>     [38,]  38 360
#>     [39,]  39 360
#>     [40,]  40 360
#>     [41,]  41 360
#>     [42,]  42 360
#>     [43,]  43 360
#>     [44,]  44 360
#>     [45,]  45 360
#>     [46,]  46 360
#>     [47,]  47 360
#>     [48,]  48 360
#>     [49,]  49 360
#>     [50,]  50 360
#>     [51,]  51 360
#>     [52,]  52 360
#>     [53,]  53 360
#>     [54,]  54 360
#>     [55,]  55 360
#>     [56,]  56 360
#>     [57,]  57 360
#>     [58,]  58 360
#>     [59,]  59 360
#>     [60,]  60 360
#>     [61,]  61 360
#>     [62,]  62 360
#>     [63,]  63 360
#>     [64,]  64 360
#>     [65,]  65 360
#>     [66,]  66 360
#>     [67,]  67 360
#>     [68,]  68 360
#>     [69,]  69 360
#>     [70,]  70 360
#>     [71,]  71 360
#>     [72,]  72 360
#>     [73,]  73 360
#>     [74,]  74 360
#>     [75,]  75 360
#>     [76,]  76 360
#>     [77,]  77 360
#>     [78,]  78 360
#>     [79,]  79 360
#>     [80,]  80 360
#>     [81,]  81 360
#>     [82,]  82 360
#>     [83,]  83 360
#>     [84,]  84 360
#>     [85,]  85 360
#>     [86,]  86 360
#>     [87,]  87 360
#>     [88,]  88 360
#>     [89,]  89 360
#>     [90,]  90 360
#>     [91,]  91 360
#>     [92,]  92 360
#>     [93,]  93 360
#>     [94,]  94 360
#>     [95,]  95 360
#>     [96,]  96 360
#>     [97,]  97 360
#>     [98,]  98 360
#>     [99,]  99 360
#>    [100,] 100 360
#>    [101,] 101 360
#>    [102,] 102 360
#>    [103,] 103 360
#>    [104,] 104 360
#>    [105,] 105 360
#>    [106,] 106 360
#>    [107,] 107 360
#>    [108,] 108 360
#>    [109,] 109 360
#>    [110,] 110 360
#>    [111,] 111 360
#>    [112,] 112 360
#>    [113,] 113 360
#>    [114,] 114 360
#>    [115,] 115 360
#>    [116,] 116 360
#>    [117,] 117 360
#>    [118,] 118 360
#>    [119,] 119 360
#>    [120,] 120 360
#>    [121,] 121 360
#>    [122,] 122 360
#>    [123,] 123 360
#>    [124,] 124 360
#>    [125,] 125 360
#>    [126,] 126 360
#>    [127,] 127 360
#>    [128,] 128 360
#>    [129,] 129 360
#>    [130,] 130 360
#>    [131,] 131 360
#>    [132,] 132 360
#>    [133,] 133 360
#>    [134,] 134 360
#>    [135,] 135 360
#>    [136,] 136 360
#>    [137,] 137 360
#>    [138,] 138 360
#>    [139,] 139 360
#>    [140,] 140 360
#>    [141,] 141 360
#>    [142,] 142 360
#>    [143,] 143 360
#>    [144,] 144 360
#>    [145,] 145 360
#>    [146,] 146 360
#>    [147,] 147 360
#>    [148,] 148 360
#>    [149,] 149 360
#>    [150,] 150 360
#>    [151,] 151 360
#>    [152,] 152 360
#>    [153,] 153 360
#>    [154,] 154 360
#>    [155,] 155 360
#>    [156,] 156 360
#>    [157,] 157 360
#>    [158,] 158 360
#>    [159,] 159 360
#>    [160,] 160 360
#>    [161,] 161 360
#>    [162,] 162 360
#>    [163,] 163 360
#>    [164,] 164 360
#>    [165,] 165 360
#>    [166,] 166 360
#>    [167,] 167 360
#>    [168,] 168 360
#>    [169,] 169 360
#>    [170,] 170 360
#>    [171,] 171 360
#>    [172,] 172 360
#>    [173,] 173 360
#>    [174,] 174 360
#>    [175,] 175 360
#>    [176,] 176 360
#>    [177,] 177 360
#>    [178,] 178 360
#>    [179,] 179 360
#>    [180,] 180 360
#>    [181,] 181 360
#>    [182,] 182 360
#>    [183,] 183 360
#>    [184,] 184 360
#>    [185,] 185 360
#>    [186,] 186 360
#>    [187,] 187 360
#>    [188,] 188 360
#>    [189,] 189 360
#>    [190,] 190 360
#>    [191,] 191 360
#>    [192,] 192 360
#>    [193,] 193 360
#>    [194,] 194 360
#>    [195,] 195 360
#>    [196,] 196 360
#>    [197,] 197 360
#>    [198,] 198 360
#>    [199,] 199 360
#>    [200,] 200 360
#>    [201,] 201 360
#>    [202,] 202 360
#>    [203,] 203 360
#>    [204,] 204 360
#>    [205,] 205 360
#>    [206,] 206 360
#>    [207,] 207 360
#>    [208,] 208 360
#>    [209,] 209 360
#>    [210,] 210 360
#>    [211,] 211 360
#>    [212,] 212 360
#>    [213,] 213 360
#>    [214,] 214 360
#>    [215,] 215 360
#>    [216,] 216 360
#>    [217,] 217 360
#>    [218,] 218 360
#>    [219,] 219 360
#>    [220,] 220 360
#>    [221,] 221 360
#>    [222,] 222 360
#>    [223,] 223 360
#>    [224,] 224 360
#>    [225,] 225 360
#>    [226,] 226 360
#>    [227,] 227 360
#>    [228,] 228 360
#>    [229,] 229 360
#>    [230,] 230 360
#>    [231,] 231 360
#>    [232,] 232 360
#>    [233,] 233 360
#>    [234,] 234 360
#>    [235,] 235 360
#>    [236,] 236 360
#>    [237,] 237 360
#>    [238,] 238 360
#>    [239,] 239 360
#>    [240,] 240 360
#>    [241,] 241 360
#>    [242,] 242 360
#>    [243,] 243 360
#>    [244,] 244 360
#>    [245,] 245 360
#>    [246,] 246 360
#>    [247,] 247 360
#>    [248,] 248 360
#>    [249,] 249 360
#>    [250,] 250 360
#>    [251,] 251 360
#>    [252,] 252 360
#>    [253,] 253 360
#>    [254,] 254 360
#>    [255,] 255 360
#>    [256,] 256 360
#>    [257,] 257 360
#>    [258,] 258 360
#>    [259,] 259 360
#>    [260,] 260 360
#>    [261,] 261 360
#>    [262,] 262 360
#>    [263,] 263 360
#>    [264,] 264 360
#>    [265,] 265 360
#>    [266,] 266 360
#>    [267,] 267 360
#>    [268,] 268 360
#>    [269,] 269 360
#>    [270,] 270 360
#>    [271,] 271 360
#>    [272,] 272 360
#>    [273,] 273 360
#>    [274,] 274 360
#>    [275,] 275 360
#>    [276,] 276 360
#>    [277,] 277 360
#>    [278,] 278 360
#>    [279,] 279 360
#>    [280,] 280 360
#>    [281,] 281 360
#>    [282,] 282 360
#>    [283,] 283 360
#>    [284,] 284 360
#>    [285,] 285 360
#>    [286,] 286 360
#>    [287,] 287 360
#>    [288,] 288 360
#>    [289,] 289 360
#>    [290,] 290 360
#>    [291,] 291 360
#>    [292,] 292 360
#>    [293,] 293 360
#>    [294,] 294 360
#>    [295,] 295 360
#>    [296,] 296 360
#>    [297,] 297 360
#>    [298,] 298 360
#>    [299,] 299 360
#>    [300,] 300 360
#>    [301,] 301 360
#>    [302,] 302 360
#>    [303,] 303 360
#>    [304,] 304 360
#>    [305,] 305 360
#>    [306,] 306 360
#>    [307,] 307 360
#>    [308,] 308 360
#>    [309,] 309 360
#>    [310,] 310 360
#>    [311,] 311 360
#>    [312,] 312 360
#>    [313,] 313 360
#>    [314,] 314 360
#>    [315,] 315 360
#>    [316,] 316 360
#>    [317,] 317 360
#>    [318,] 318 360
#>    [319,] 319 360
#>    [320,] 320 360
#>    [321,] 321 360
#>    [322,] 322 360
#>    [323,] 323 360
#>    [324,] 324 360
#>    [325,] 325 360
#>    [326,] 326 360
#>    [327,] 327 360
#>    [328,] 328 360
#>    [329,] 329 360
#>    [330,] 330 360
#>    [331,] 331 360
#>    [332,] 332 360
#>    [333,] 333 360
#>    [334,] 334 360
#>    [335,] 335 360
#>    [336,] 336 360
#>    [337,] 337 360
#>    [338,] 338 360
#>    [339,] 339 360
#>    [340,] 340 360
#>    [341,] 341 360
#>    [342,] 342 360
#>    [343,] 343 360
#>    [344,] 344 360
#>    [345,] 345 360
#>    [346,] 346 360
#>    [347,] 347 360
#>    [348,] 348 360
#>    [349,] 349 360
#>    [350,] 350 360
#>    [351,] 351 360
#>    [352,] 352 360
#>    [353,] 353 360
#>    [354,] 354 360
#>    [355,] 355 360
#>    [356,] 356 360
#>    [357,] 357 360
#>    [358,] 358 360
#>    [359,] 359 360
#>    [360,] 360 360
#>    [361,] 361 360
#>    [362,] 362 360
#>    [363,] 363 360
#>    [364,] 364 360
#>    [365,] 365 360
#>    [366,] 366 360
#>    [367,] 367 360
#>    [368,] 368 360
#>    [369,] 369 360
#>    [370,] 370 360
#>    [371,] 371 360
#>    [372,] 372 360
#>    [373,] 373 360
#>    [374,] 374 360
#>    [375,] 375 360
#>    [376,] 376 360
#>    [377,] 377 360
#>    [378,] 378 360
#>    [379,] 379 360
#>    [380,] 380 360
#>    [381,] 381 360
#>    [382,] 382 360
#>    [383,] 383 360
#>    [384,] 384 360
#>    [385,] 385 360
#>    [386,] 386 360
#>    [387,] 387 360
#>    [388,] 388 360
#>    [389,] 389 360
#>    [390,] 390 360
#>    [391,] 391 360
#>    [392,] 392 360
#>    [393,] 393 360
#>    [394,] 394 360
#>    [395,] 395 360
#>    [396,] 396 360
#>    [397,] 397 360
#>    [398,] 398 360
#>    [399,] 399 360
#>    [400,] 400 360
#>    [401,] 401 360
#>    [402,] 402 360
#>    [403,] 403 360
#>    [404,] 404 360
#>    [405,] 405 360
#>    [406,] 406 360
#>    [407,] 407 360
#>    [408,] 408 360
#>    [409,] 409 360
#>    [410,] 410 360
#>    [411,] 411 360
#>    [412,] 412 360
#>    [413,] 413 360
#>    [414,] 414 360
#>    [415,] 415 360
#>    [416,] 416 360
#>    [417,] 417 360
#>    [418,] 418 360
#>    [419,] 419 360
#>    [420,] 420 360
#>    [421,] 421 360
#>    [422,] 422 360
#>    [423,] 423 360
#>    [424,] 424 360
#>    [425,] 425 360
#>    [426,] 426 360
#>    [427,] 427 360
#>    [428,] 428 360
#>    [429,] 429 360
#>    [430,] 430 360
#>    [431,] 431 360
#>    [432,] 432 360
#>    [433,] 433 360
#>    [434,] 434 360
#>    [435,] 435 360
#>    [436,] 436 360
#>    [437,] 437 360
#>    [438,] 438 360
#>    [439,] 439 360
#>    [440,] 440 360
#>    [441,] 441 360
#>    [442,] 442 360
#>    [443,] 443 360
#>    [444,] 444 360
#>    [445,] 445 360
#>    [446,] 446 360
#>    [447,] 447 360
#>    [448,] 448 360
#>    [449,] 449 360
#>    [450,] 450 360
#>    [451,] 451 360
#>    [452,] 452 360
#>    [453,] 453 360
#>    [454,] 454 360
#>    [455,] 455 360
#>    [456,] 456 360
#>    [457,] 457 360
#>    [458,] 458 360
#>    [459,] 459 360
#>    [460,] 460 360
#>    [461,] 461 360
#>    [462,] 462 360
#>    [463,] 463 360
#>    [464,] 464 360
#>    [465,] 465 360
#>    [466,] 466 360
#>    [467,] 467 360
#>    [468,] 468 360
#>    [469,] 469 360
#>    [470,] 470 360
#>    [471,] 471 360
#>    [472,] 472 360
#>    [473,] 473 360
#>    [474,] 474 360
#>    [475,] 475 360
#>    [476,] 476 360
#>    [477,] 477 360
#>    [478,] 478 360
#>    [479,] 479 360
#>    [480,] 480 360
#>    [481,] 481 360
#>    [482,] 482 360
#>    [483,] 483 360
#>    [484,] 484 360
#>    [485,] 485 360
#>    [486,] 486 360
#>    [487,] 487 360
#>    [488,] 488 360
#>    [489,] 489 360
#>    [490,] 490 360
#>    [491,] 491 360
#>    [492,] 492 360
#>    [493,] 493 360
#>    [494,] 494 360
#>    [495,] 495 360
#>    [496,] 496 360
#>    [497,] 497 360
#>    [498,] 498 360
#>    [499,] 499 360
#>    [500,] 500 360
#>  [ reached getOption("max.print") -- omitted 229900 rows ]