Calculate the Height and Width of a Text String

getTextSize calculates the size of a box that contains the specified text string, the tight box surrounding it, and the baseline.

Usage

getTextSize(
  text,
  font_face = "simplex",
  font_scale = 1,
  italic = FALSE,
  thickness = 1
)

Arguments

text

A character string representing the text to be drawn.

font_face

A character string representing the font type of the text (default: "simplex"). See notes for a list of available font types.

font_scale

A numeric value representing the scale factor by which the font-specific base size is multiplied (default: 1).

italic

A logical specifying whether the text should italicized (default: FALSE).

thickness

A numeric value representing the thickness in pixels of the line (default: 1).

Value

A two-element vector corresponding to the height and width of the text string.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

getTextSize("I'm a balloon")
#> [1]  22 209