Histogram Matching/Specification

histmatch transforms an Image object so that its histogram matches (approximately) that of another Image object.

histmatch(image, reference, target = "new")

Arguments

image

An Image object to transform.

reference

An Image object which histogram will be used as a reference to transform image.

target

The location where the results should be stored. It can take 3 values:

"new":a new Image object is created and the results are stored inside (the default).

"self":the results are stored back into image (faster but destructive).

An Image object:the results are stored in another existing Image object. This is fast and will not replace the content of image but will replace that of target. Note that if target does not have the same dimensions, number of channels, and bit depth as image, nothing will be stored.

Value

If target="new", the function returns an Image object. If target="self", the function returns nothing and modifies image in place. If target is an Image object, the function returns nothing and modifies that Image object in place.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
dots <- image(system.file("sample_img/dots.jpg", package = "Rvision"))
dots_matched <- histmatch(dots, balloon)