Pixel Values Along a Line Segment

improfile finds all pixels intersected by a line segment and returns their coordinates and color values.

Usage

improfile(image, xi, yi, connectivity = 8, left_to_right = FALSE)

Arguments

image

An Image object.

xi, yi

Two-element vectors containing the x and y coordinates of the start and end points of a line segment. The points can lie outside of the image boundaries but then the line will be clipped on the image boundaries.

connectivity

The connectivity neighborhood to decide whether 2 pixels are contiguous. This parameter can take two values:

4:

the neighborhood of a pixel are the four pixels located above (north), below (south), to the left (west) and right (east) of the pixel.

8 (the default):

the neighborhood of a pixel includes the four 4-neighbors and the four pixels along the diagonal directions (northeast, northwest, southeast, and southwest).

left_to_right

If `TRUE`, the line segment is traversed from the leftmost endpoint to the rightmost endpoint, regardless of the order in which they are specified in xi and yi. Otherwise (the default), the line is traversed in the order in which the points are specified in xi and yi.

Value

A matrix in which the first two columns indicate the coordinates of the points traversed by the line segment, and the other columns indicate the color values at each location.

Author

Simon Garnier, garnier@njit.edu

Examples

dots <- image(system.file("sample_img/dots.jpg", package = "Rvision"))
improfile(dots, c(1, ncol(dots)), c(nrow(dots) / 2, nrow(dots) / 2))
#>            X   Y   B   G   R
#>    [1,]    1 640 255 255 255
#>    [2,]    2 640 255 255 255
#>    [3,]    3 640 255 255 255
#>    [4,]    4 640 255 255 255
#>    [5,]    5 640 255 255 255
#>    [6,]    6 640 255 255 255
#>    [7,]    7 640 255 255 255
#>    [8,]    8 640 255 255 255
#>    [9,]    9 640 255 255 255
#>   [10,]   10 640 255 255 255
#>   [11,]   11 640 255 255 255
#>   [12,]   12 640 255 255 255
#>   [13,]   13 640 255 255 255
#>   [14,]   14 640 255 255 255
#>   [15,]   15 640 255 255 255
#>   [16,]   16 640 255 255 255
#>   [17,]   17 640 255 255 255
#>   [18,]   18 640 255 255 255
#>   [19,]   19 640 255 255 255
#>   [20,]   20 640 255 255 255
#>   [21,]   21 640 255 255 255
#>   [22,]   22 640 255 255 255
#>   [23,]   23 640 255 255 255
#>   [24,]   24 640 255 255 255
#>   [25,]   25 640 255 255 255
#>   [26,]   26 640 255 255 255
#>   [27,]   27 640 255 255 255
#>   [28,]   28 640 255 255 255
#>   [29,]   29 640 255 255 255
#>   [30,]   30 640 255 255 255
#>   [31,]   31 640 255 255 255
#>   [32,]   32 640 255 255 255
#>   [33,]   33 640 255 255 255
#>   [34,]   34 640 255 255 255
#>   [35,]   35 640 255 255 255
#>   [36,]   36 640 255 255 255
#>   [37,]   37 640 255 255 255
#>   [38,]   38 640 255 255 255
#>   [39,]   39 640 255 255 255
#>   [40,]   40 640 255 255 255
#>   [41,]   41 640 255 255 255
#>   [42,]   42 640 255 255 255
#>   [43,]   43 640 255 255 255
#>   [44,]   44 640 255 255 255
#>   [45,]   45 640 255 255 255
#>   [46,]   46 640 255 255 255
#>   [47,]   47 640 255 255 255
#>   [48,]   48 640 255 255 255
#>   [49,]   49 640 255 255 255
#>   [50,]   50 640 255 255 255
#>   [51,]   51 640 255 255 255
#>   [52,]   52 640 255 255 255
#>   [53,]   53 640 255 255 255
#>   [54,]   54 640 255 255 255
#>   [55,]   55 640 255 255 255
#>   [56,]   56 640 255 255 255
#>   [57,]   57 640 255 255 255
#>   [58,]   58 640 255 255 255
#>   [59,]   59 640 255 255 255
#>   [60,]   60 640 255 255 255
#>   [61,]   61 640 255 255 255
#>   [62,]   62 640 255 255 255
#>   [63,]   63 640 255 255 255
#>   [64,]   64 640 255 255 255
#>   [65,]   65 640 255 255 255
#>   [66,]   66 640 255 255 255
#>   [67,]   67 640 255 255 255
#>   [68,]   68 640 255 255 255
#>   [69,]   69 640 255 255 255
#>   [70,]   70 640 255 255 255
#>   [71,]   71 640 255 255 255
#>   [72,]   72 640 255 255 255
#>   [73,]   73 640 255 255 255
#>   [74,]   74 640 255 255 255
#>   [75,]   75 640 255 255 255
#>   [76,]   76 640 255 255 255
#>   [77,]   77 640 255 255 255
#>   [78,]   78 640 255 255 255
#>   [79,]   79 640 254 254 254
#>   [80,]   80 640 255 255 255
#>   [81,]   81 640 255 255 255
#>   [82,]   82 640 255 255 255
#>   [83,]   83 640 255 255 255
#>   [84,]   84 640 255 255 255
#>   [85,]   85 640 255 255 255
#>   [86,]   86 640 255 255 255
#>   [87,]   87 640 255 255 255
#>   [88,]   88 640 255 255 255
#>   [89,]   89 640 255 255 255
#>   [90,]   90 640 255 255 255
#>   [91,]   91 640 255 255 255
#>   [92,]   92 640 255 255 255
#>   [93,]   93 640 255 255 255
#>   [94,]   94 640 255 255 255
#>   [95,]   95 640 255 255 255
#>   [96,]   96 640 255 255 255
#>   [97,]   97 640 255 255 255
#>   [98,]   98 640 255 255 255
#>   [99,]   99 640 255 255 255
#>  [100,]  100 640 255 255 255
#>  [101,]  101 640 255 255 255
#>  [102,]  102 640 255 255 255
#>  [103,]  103 640 255 255 255
#>  [104,]  104 640 255 255 255
#>  [105,]  105 640 255 255 255
#>  [106,]  106 640 255 255 255
#>  [107,]  107 640 255 255 255
#>  [108,]  108 640 255 255 255
#>  [109,]  109 640 255 255 255
#>  [110,]  110 640 255 255 255
#>  [111,]  111 640 255 255 255
#>  [112,]  112 640 255 255 255
#>  [113,]  113 640 255 255 255
#>  [114,]  114 640 255 255 255
#>  [115,]  115 640 255 255 255
#>  [116,]  116 640 255 255 255
#>  [117,]  117 640 255 255 255
#>  [118,]  118 640 255 255 255
#>  [119,]  119 640 255 255 255
#>  [120,]  120 640 255 255 255
#>  [121,]  121 640 255 255 255
#>  [122,]  122 640 255 255 255
#>  [123,]  123 640 255 255 255
#>  [124,]  124 640 255 255 255
#>  [125,]  125 640 255 255 255
#>  [126,]  126 640 255 255 255
#>  [127,]  127 640 255 255 255
#>  [128,]  128 640 255 255 255
#>  [129,]  129 640 255 255 255
#>  [130,]  130 640 255 255 255
#>  [131,]  131 640 255 255 255
#>  [132,]  132 640 255 255 255
#>  [133,]  133 640 255 255 255
#>  [134,]  134 640 255 255 255
#>  [135,]  135 640 255 255 255
#>  [136,]  136 640 255 255 255
#>  [137,]  137 640 255 255 255
#>  [138,]  138 640 255 255 255
#>  [139,]  139 640 255 255 255
#>  [140,]  140 640 255 255 255
#>  [141,]  141 640 255 255 255
#>  [142,]  142 640 255 255 255
#>  [143,]  143 640 255 255 255
#>  [144,]  144 640 255 255 255
#>  [145,]  145 640 250 254 255
#>  [146,]  146 640 251 255 255
#>  [147,]  147 640 253 255 255
#>  [148,]  148 640 253 255 255
#>  [149,]  149 640 253 255 255
#>  [150,]  150 640 253 255 255
#>  [151,]  151 640 255 255 255
#>  [152,]  152 640 255 255 255
#>  [153,]  153 640 255 255 255
#>  [154,]  154 640 255 255 255
#>  [155,]  155 640 255 255 255
#>  [156,]  156 640 255 255 255
#>  [157,]  157 640 255 255 255
#>  [158,]  158 640 255 255 255
#>  [159,]  159 640 255 255 255
#>  [160,]  160 640 255 255 255
#>  [161,]  161 640 247 252 251
#>  [162,]  162 640 250 255 254
#>  [163,]  163 640 244 252 251
#>  [164,]  164 640 246 251 250
#>  [165,]  165 640 242 255 253
#>  [166,]  166 640 238 252 248
#>  [167,]  167 640 240 251 248
#>  [168,]  168 640 242 245 243
#>  [169,]  169 640 253 254 252
#>  [170,]  170 640 255 244 245
#>  [171,]  171 640 126  70  75
#>  [172,]  172 640  72   5  12
#>  [173,]  173 640 100  32  39
#>  [174,]  174 640 105  32  40
#>  [175,]  175 640  90  32  37
#>  [176,]  176 640  88  30  35
#>  [177,]  177 640  99  33  38
#>  [178,]  178 640  99  33  38
#>  [179,]  179 640  98  32  37
#>  [180,]  180 640  98  32  37
#>  [181,]  181 640  99  32  37
#>  [182,]  182 640 100  33  38
#>  [183,]  183 640 100  33  38
#>  [184,]  184 640 101  34  39
#>  [185,]  185 640 101  34  39
#>  [186,]  186 640 102  35  40
#>  [187,]  187 640 102  35  40
#>  [188,]  188 640 102  35  40
#>  [189,]  189 640 104  34  40
#>  [190,]  190 640 104  34  40
#>  [191,]  191 640 104  34  40
#>  [192,]  192 640 104  34  40
#>  [193,]  193 640 103  36  41
#>  [194,]  194 640 102  35  40
#>  [195,]  195 640 102  35  40
#>  [196,]  196 640 102  35  40
#>  [197,]  197 640 102  35  38
#>  [198,]  198 640 102  35  38
#>  [199,]  199 640 105  36  39
#>  [200,]  200 640 105  36  39
#>  [201,]  201 640 105  36  39
#>  [202,]  202 640 105  36  39
#>  [203,]  203 640 106  38  39
#>  [204,]  204 640 106  38  39
#>  [205,]  205 640 107  39  40
#>  [206,]  206 640 108  40  41
#>  [207,]  207 640 109  41  42
#>  [208,]  208 640 106  38  39
#>  [209,]  209 640 107  39  40
#>  [210,]  210 640 108  40  41
#>  [211,]  211 640 109  41  42
#>  [212,]  212 640 109  41  42
#>  [213,]  213 640 108  40  41
#>  [214,]  214 640 108  40  41
#>  [215,]  215 640 110  40  40
#>  [216,]  216 640 111  41  41
#>  [217,]  217 640 113  41  41
#>  [218,]  218 640 111  39  39
#>  [219,]  219 640 112  40  40
#>  [220,]  220 640 113  41  41
#>  [221,]  221 640 115  41  41
#>  [222,]  222 640 115  41  41
#>  [223,]  223 640 116  42  42
#>  [224,]  224 640 118  44  44
#>  [225,]  225 640 114  43  40
#>  [226,]  226 640 115  44  41
#>  [227,]  227 640 116  45  42
#>  [228,]  228 640 116  45  42
#>  [229,]  229 640 117  43  41
#>  [230,]  230 640 117  43  41
#>  [231,]  231 640 117  43  41
#>  [232,]  232 640 118  44  42
#>  [233,]  233 640 118  43  41
#>  [234,]  234 640 118  43  41
#>  [235,]  235 640 118  43  41
#>  [236,]  236 640 118  43  41
#>  [237,]  237 640 121  44  41
#>  [238,]  238 640 122  45  42
#>  [239,]  239 640 121  44  41
#>  [240,]  240 640 120  43  40
#>  [241,]  241 640 114  46  41
#>  [242,]  242 640 111  42  39
#>  [243,]  243 640 119  47  40
#>  [244,]  244 640 121  43  36
#>  [245,]  245 640 140  49  44
#>  [246,]  246 640 126  35  30
#>  [247,]  247 640  90  38  31
#>  [248,]  248 640 217 191 185
#>  [249,]  249 640 252 255 251
#>  [250,]  250 640 237 245 244
#>  [251,]  251 640 254 250 255
#>  [252,]  252 640 255 251 253
#>  [253,]  253 640 255 250 255
#>  [254,]  254 640 255 254 254
#>  [255,]  255 640 242 254 254
#>  [256,]  256 640 242 255 253
#>  [257,]  257 640 254 254 254
#>  [258,]  258 640 254 254 254
#>  [259,]  259 640 254 254 254
#>  [260,]  260 640 255 255 255
#>  [261,]  261 640 255 255 255
#>  [262,]  262 640 255 255 255
#>  [263,]  263 640 255 255 255
#>  [264,]  264 640 255 255 255
#>  [265,]  265 640 255 255 255
#>  [266,]  266 640 255 255 255
#>  [267,]  267 640 255 255 255
#>  [268,]  268 640 255 255 255
#>  [269,]  269 640 255 255 255
#>  [270,]  270 640 254 254 254
#>  [271,]  271 640 254 254 254
#>  [272,]  272 640 254 254 254
#>  [273,]  273 640 254 254 254
#>  [274,]  274 640 254 254 254
#>  [275,]  275 640 254 254 254
#>  [276,]  276 640 255 255 255
#>  [277,]  277 640 255 255 255
#>  [278,]  278 640 255 255 255
#>  [279,]  279 640 255 255 255
#>  [280,]  280 640 255 255 255
#>  [281,]  281 640 255 255 255
#>  [282,]  282 640 255 255 255
#>  [283,]  283 640 255 255 255
#>  [284,]  284 640 255 255 255
#>  [285,]  285 640 255 255 255
#>  [286,]  286 640 254 254 254
#>  [287,]  287 640 254 254 254
#>  [288,]  288 640 254 254 254
#>  [289,]  289 640 254 254 254
#>  [290,]  290 640 254 254 254
#>  [291,]  291 640 254 254 254
#>  [292,]  292 640 255 255 255
#>  [293,]  293 640 255 255 255
#>  [294,]  294 640 255 255 255
#>  [295,]  295 640 255 255 255
#>  [296,]  296 640 255 255 255
#>  [297,]  297 640 252 252 252
#>  [298,]  298 640 255 255 255
#>  [299,]  299 640 251 251 251
#>  [300,]  300 640 255 255 255
#>  [301,]  301 640 251 251 251
#>  [302,]  302 640 251 249 249
#>  [303,]  303 640 255 255 255
#>  [304,]  304 640 159 157 157
#>  [305,]  305 640  95  32  18
#>  [306,]  306 640 139  67  49
#>  [307,]  307 640 132  60  43
#>  [308,]  308 640 139  56  41
#>  [309,]  309 640 144  55  34
#>  [310,]  310 640 154  63  42
#>  [311,]  311 640 145  53  36
#>  [312,]  312 640 156  60  44
#>  [313,]  313 640 142  62  45
#>  [314,]  314 640 143  61  44
#>  [315,]  315 640 142  63  44
#>  [316,]  316 640 143  63  46
#>  [317,]  317 640 154  62  43
#>  [318,]  318 640 156  61  45
#>  [319,]  319 640 158  60  44
#>  [320,]  320 640 157  59  43
#>  [321,]  321 640 154  63  42
#>  [322,]  322 640 153  62  41
#>  [323,]  323 640 150  61  40
#>  [324,]  324 640 152  63  42
#>  [325,]  325 640 150  64  40
#>  [326,]  326 640 152  64  40
#>  [327,]  327 640 156  64  39
#>  [328,]  328 640 158  65  40
#>  [329,]  329 640 157  64  38
#>  [330,]  330 640 157  64  38
#>  [331,]  331 640 158  65  39
#>  [332,]  332 640 157  66  39
#>  [333,]  333 640 156  67  40
#>  [334,]  334 640 152  68  40
#>  [335,]  335 640 152  69  41
#>  [336,]  336 640 152  69  41
#>  [337,]  337 640 155  68  42
#>  [338,]  338 640 155  68  42
#>  [339,]  339 640 155  68  42
#>  [340,]  340 640 154  69  43
#>  [341,]  341 640 154  69  43
#>  [342,]  342 640 154  69  43
#>  [343,]  343 640 154  68  44
#>  [344,]  344 640 154  68  44
#>  [345,]  345 640 155  70  44
#>  [346,]  346 640 155  70  44
#>  [347,]  347 640 157  72  46
#>  [348,]  348 640 156  72  46
#>  [349,]  349 640 156  73  45
#>  [350,]  350 640 156  73  45
#>  [351,]  351 640 156  73  45
#>  [352,]  352 640 156  73  45
#>  [353,]  353 640 163  71  46
#>  [354,]  354 640 164  72  47
#>  [355,]  355 640 167  74  49
#>  [356,]  356 640 166  74  49
#>  [357,]  357 640 162  75  49
#>  [358,]  358 640 159  75  49
#>  [359,]  359 640 157  76  49
#>  [360,]  360 640 157  76  49
#>  [361,]  361 640 153  77  48
#>  [362,]  362 640 156  78  49
#>  [363,]  363 640 156  79  47
#>  [364,]  364 640 158  80  44
#>  [365,]  365 640 156  81  43
#>  [366,]  366 640 160  80  43
#>  [367,]  367 640 160  80  43
#>  [368,]  368 640 161  81  44
#>  [369,]  369 640 162  75  49
#>  [370,]  370 640 167  76  49
#>  [371,]  371 640 158  81  48
#>  [372,]  372 640 154  85  46
#>  [373,]  373 640 155  85  45
#>  [374,]  374 640 153  84  44
#>  [375,]  375 640 162  83  44
#>  [376,]  376 640 170  82  46
#>  [377,]  377 640 168  78  47
#>  [378,]  378 640 176  84  59
#>  [379,]  379 640 167  80  54
#>  [380,]  380 640 157  73  48
#>  [381,]  381 640 162  83  56
#>  [382,]  382 640 156  87  54
#>  [383,]  383 640 153  86  47
#>  [384,]  384 640 136  74  34
#>  [385,]  385 640 120 117 113
#>  [386,]  386 640 253 250 246
#>  [387,]  387 640 255 255 251
#>  [388,]  388 640 248 247 243
#>  [389,]  389 640 250 251 247
#>  [390,]  390 640 254 255 253
#>  [391,]  391 640 252 255 253
#>  [392,]  392 640 252 254 254
#>  [393,]  393 640 252 255 253
#>  [394,]  394 640 252 255 253
#>  [395,]  395 640 252 254 254
#>  [396,]  396 640 252 254 254
#>  [397,]  397 640 254 254 254
#>  [398,]  398 640 255 254 254
#>  [399,]  399 640 255 254 254
#>  [400,]  400 640 255 254 254
#>  [401,]  401 640 254 254 254
#>  [402,]  402 640 254 254 254
#>  [403,]  403 640 254 254 254
#>  [404,]  404 640 254 254 254
#>  [405,]  405 640 254 254 254
#>  [406,]  406 640 255 255 255
#>  [407,]  407 640 255 255 255
#>  [408,]  408 640 255 255 255
#>  [409,]  409 640 255 255 255
#>  [410,]  410 640 255 255 255
#>  [411,]  411 640 255 255 255
#>  [412,]  412 640 254 254 254
#>  [413,]  413 640 254 254 254
#>  [414,]  414 640 254 254 254
#>  [415,]  415 640 254 254 254
#>  [416,]  416 640 254 254 254
#>  [417,]  417 640 254 254 254
#>  [418,]  418 640 254 254 254
#>  [419,]  419 640 254 254 254
#>  [420,]  420 640 254 254 254
#>  [421,]  421 640 254 254 254
#>  [422,]  422 640 255 255 255
#>  [423,]  423 640 255 255 255
#>  [424,]  424 640 255 255 255
#>  [425,]  425 640 255 255 255
#>  [426,]  426 640 255 255 255
#>  [427,]  427 640 255 255 255
#>  [428,]  428 640 254 254 254
#>  [429,]  429 640 254 254 254
#>  [430,]  430 640 254 254 254
#>  [431,]  431 640 254 254 254
#>  [432,]  432 640 254 254 254
#>  [433,]  433 640 254 254 254
#>  [434,]  434 640 254 254 254
#>  [435,]  435 640 254 254 254
#>  [436,]  436 640 254 254 254
#>  [437,]  437 640 254 254 254
#>  [438,]  438 640 255 255 255
#>  [439,]  439 640 255 255 255
#>  [440,]  440 640 255 255 255
#>  [441,]  441 640 255 255 255
#>  [442,]  442 640 255 255 255
#>  [443,]  443 640 255 255 255
#>  [444,]  444 640 254 254 254
#>  [445,]  445 640 254 254 254
#>  [446,]  446 640 254 254 254
#>  [447,]  447 640 254 254 254
#>  [448,]  448 640 254 254 254
#>  [449,]  449 640 254 254 254
#>  [450,]  450 640 254 254 254
#>  [451,]  451 640 254 254 254
#>  [452,]  452 640 254 254 254
#>  [453,]  453 640 254 254 254
#>  [454,]  454 640 255 255 255
#>  [455,]  455 640 255 255 255
#>  [456,]  456 640 255 255 255
#>  [457,]  457 640 255 255 255
#>  [458,]  458 640 255 255 255
#>  [459,]  459 640 255 255 255
#>  [460,]  460 640 254 254 254
#>  [461,]  461 640 254 254 254
#>  [462,]  462 640 254 254 254
#>  [463,]  463 640 254 254 254
#>  [464,]  464 640 254 254 254
#>  [465,]  465 640 254 254 254
#>  [466,]  466 640 254 254 254
#>  [467,]  467 640 254 254 254
#>  [468,]  468 640 254 254 254
#>  [469,]  469 640 254 254 254
#>  [470,]  470 640 255 255 255
#>  [471,]  471 640 255 255 255
#>  [472,]  472 640 255 255 255
#>  [473,]  473 640 255 255 255
#>  [474,]  474 640 255 255 255
#>  [475,]  475 640 255 255 255
#>  [476,]  476 640 254 254 254
#>  [477,]  477 640 254 254 254
#>  [478,]  478 640 254 254 254
#>  [479,]  479 640 254 254 254
#>  [480,]  480 640 254 254 254
#>  [481,]  481 640 255 255 255
#>  [482,]  482 640 254 253 255
#>  [483,]  483 640 245 255 254
#>  [484,]  484 640 245 255 255
#>  [485,]  485 640 254 253 255
#>  [486,]  486 640 255 254 254
#>  [487,]  487 640 253 253 253
#>  [488,]  488 640 255 253 253
#>  [489,]  489 640 247 254 251
#>  [490,]  490 640 249 254 252
#>  [491,]  491 640 252 255 251
#>  [492,]  492 640 249 254 253
#>  [493,]  493 640 248 253 254
#>  [494,]  494 640 248 252 253
#>  [495,]  495 640 233 250 253
#>  [496,]  496 640 241 255 254
#>  [497,]  497 640 251 249 239
#>  [498,]  498 640 253 255 234
#>  [499,]  499 640 254 237 210
#>  [500,]  500 640 175 130  96
#>  [501,]  501 640 168 108  62
#>  [502,]  502 640 203 131  77
#>  [503,]  503 640 205 129  70
#>  [504,]  504 640 206 128  69
#>  [505,]  505 640 196 130  71
#>  [506,]  506 640 192 129  71
#>  [507,]  507 640 187 130  75
#>  [508,]  508 640 187 132  75
#>  [509,]  509 640 187 131  72
#>  [510,]  510 640 193 130  72
#>  [511,]  511 640 194 131  70
#>  [512,]  512 640 198 133  72
#>  [513,]  513 640 193 134  72
#>  [514,]  514 640 193 134  72
#>  [515,]  515 640 193 134  72
#>  [516,]  516 640 193 134  72
#>  [517,]  517 640 193 135  70
#>  [518,]  518 640 194 136  71
#>  [519,]  519 640 196 138  73
#>  [520,]  520 640 197 139  74
#>  [521,]  521 640 195 137  72
#>  [522,]  522 640 195 137  72
#>  [523,]  523 640 194 139  72
#>  [524,]  524 640 194 139  72
#>  [525,]  525 640 195 140  73
#>  [526,]  526 640 195 140  73
#>  [527,]  527 640 195 140  73
#>  [528,]  528 640 195 140  73
#>  [529,]  529 640 194 141  74
#>  [530,]  530 640 195 142  75
#>  [531,]  531 640 195 142  75
#>  [532,]  532 640 196 143  76
#>  [533,]  533 640 196 143  76
#>  [534,]  534 640 197 144  77
#>  [535,]  535 640 200 145  78
#>  [536,]  536 640 200 145  78
#>  [537,]  537 640 200 145  78
#>  [538,]  538 640 200 145  78
#>  [539,]  539 640 200 145  78
#>  [540,]  540 640 201 146  79
#>  [541,]  541 640 201 146  79
#>  [542,]  542 640 201 146  79
#>  [543,]  543 640 202 147  80
#>  [544,]  544 640 202 147  80
#>  [545,]  545 640 200 148  78
#>  [546,]  546 640 200 148  78
#>  [547,]  547 640 201 149  79
#>  [548,]  548 640 202 150  80
#>  [549,]  549 640 203 151  80
#>  [550,]  550 640 203 151  80
#>  [551,]  551 640 203 151  80
#>  [552,]  552 640 203 151  80
#>  [553,]  553 640 204 154  82
#>  [554,]  554 640 203 153  81
#>  [555,]  555 640 203 154  80
#>  [556,]  556 640 203 154  80
#>  [557,]  557 640 203 154  80
#>  [558,]  558 640 204 155  81
#>  [559,]  559 640 206 157  83
#>  [560,]  560 640 206 157  83
#>  [561,]  561 640 205 158  72
#>  [562,]  562 640 205 159  75
#>  [563,]  563 640 206 158  80
#>  [564,]  564 640 205 157  85
#>  [565,]  565 640 205 155  89
#>  [566,]  566 640 204 156  90
#>  [567,]  567 640 206 157  87
#>  [568,]  568 640 209 158  85
#>  [569,]  569 640 209 161  79
#>  [570,]  570 640 212 165  81
#>  [571,]  571 640 209 169  94
#>  [572,]  572 640 179 147  74
#>  [573,]  573 640 192 163 106
#>  [574,]  574 640 249 230 187
#>  [575,]  575 640 255 255 226
#>  [576,]  576 640 255 251 233
#>  [577,]  577 640 240 246 251
#>  [578,]  578 640 248 253 254
#>  [579,]  579 640 251 250 246
#>  [580,]  580 640 255 254 253
#>  [581,]  581 640 243 251 251
#>  [582,]  582 640 248 255 255
#>  [583,]  583 640 244 249 252
#>  [584,]  584 640 249 253 254
#>  [585,]  585 640 254 255 246
#>  [586,]  586 640 255 255 246
#>  [587,]  587 640 255 254 250
#>  [588,]  588 640 254 255 253
#>  [589,]  589 640 249 253 254
#>  [590,]  590 640 249 252 255
#>  [591,]  591 640 252 255 251
#>  [592,]  592 640 255 255 251
#>  [593,]  593 640 254 254 254
#>  [594,]  594 640 254 254 254
#>  [595,]  595 640 254 254 254
#>  [596,]  596 640 254 254 254
#>  [597,]  597 640 254 254 254
#>  [598,]  598 640 255 255 255
#>  [599,]  599 640 255 255 255
#>  [600,]  600 640 255 255 255
#>  [601,]  601 640 255 255 255
#>  [602,]  602 640 255 255 255
#>  [603,]  603 640 255 255 255
#>  [604,]  604 640 254 254 254
#>  [605,]  605 640 254 254 254
#>  [606,]  606 640 254 254 254
#>  [607,]  607 640 254 254 254
#>  [608,]  608 640 254 254 254
#>  [609,]  609 640 254 254 254
#>  [610,]  610 640 254 254 254
#>  [611,]  611 640 254 254 254
#>  [612,]  612 640 254 254 254
#>  [613,]  613 640 254 254 254
#>  [614,]  614 640 255 255 255
#>  [615,]  615 640 255 255 255
#>  [616,]  616 640 255 255 255
#>  [617,]  617 640 255 255 255
#>  [618,]  618 640 255 255 255
#>  [619,]  619 640 255 255 255
#>  [620,]  620 640 254 254 254
#>  [621,]  621 640 254 254 254
#>  [622,]  622 640 254 254 254
#>  [623,]  623 640 254 254 254
#>  [624,]  624 640 254 254 254
#>  [625,]  625 640 253 252 254
#>  [626,]  626 640 253 252 255
#>  [627,]  627 640 253 255 243
#>  [628,]  628 640 255 255 243
#>  [629,]  629 640 252 254 254
#>  [630,]  630 640 250 252 253
#>  [631,]  631 640 254 251 247
#>  [632,]  632 640 253 252 238
#>  [633,]  633 640 211 202 138
#>  [634,]  634 640 213 198 106
#>  [635,]  635 640 234 205  97
#>  [636,]  636 640 228 200  92
#>  [637,]  637 640 226 197  99
#>  [638,]  638 640 235 200 110
#>  [639,]  639 640 212 200  98
#>  [640,]  640 640 218 204 102
#>  [641,]  641 640 226 203 104
#>  [642,]  642 640 224 203 104
#>  [643,]  643 640 223 202 103
#>  [644,]  644 640 221 200 101
#>  [645,]  645 640 222 201 102
#>  [646,]  646 640 224 201 102
#>  [647,]  647 640 222 198 102
#>  [648,]  648 640 219 195  99
#>  [649,]  649 640 221 197 101
#>  [650,]  650 640 221 197 101
#>  [651,]  651 640 221 196 102
#>  [652,]  652 640 221 195 101
#>  [653,]  653 640 221 195 101
#>  [654,]  654 640 219 193  99
#>  [655,]  655 640 220 194 100
#>  [656,]  656 640 220 194 100
#>  [657,]  657 640 222 193 102
#>  [658,]  658 640 221 192 101
#>  [659,]  659 640 220 191 100
#>  [660,]  660 640 220 191 100
#>  [661,]  661 640 220 191 100
#>  [662,]  662 640 220 191 100
#>  [663,]  663 640 221 189 100
#>  [664,]  664 640 220 188  99
#>  [665,]  665 640 219 187  98
#>  [666,]  666 640 219 187  98
#>  [667,]  667 640 218 186  97
#>  [668,]  668 640 218 186  97
#>  [669,]  669 640 219 184  98
#>  [670,]  670 640 219 184  98
#>  [671,]  671 640 218 183  97
#>  [672,]  672 640 218 183  97
#>  [673,]  673 640 215 183  94
#>  [674,]  674 640 215 183  94
#>  [675,]  675 640 215 182  96
#>  [676,]  676 640 214 181  95
#>  [677,]  677 640 214 181  95
#>  [678,]  678 640 213 180  94
#>  [679,]  679 640 212 179  93
#>  [680,]  680 640 212 179  93
#>  [681,]  681 640 211 177  93
#>  [682,]  682 640 211 177  93
#>  [683,]  683 640 213 177  93
#>  [684,]  684 640 213 177  93
#>  [685,]  685 640 214 177  93
#>  [686,]  686 640 212 174  92
#>  [687,]  687 640 213 175  93
#>  [688,]  688 640 212 174  92
#>  [689,]  689 640 210 173  89
#>  [690,]  690 640 212 175  91
#>  [691,]  691 640 213 176  92
#>  [692,]  692 640 211 174  90
#>  [693,]  693 640 210 173  89
#>  [694,]  694 640 210 173  89
#>  [695,]  695 640 211 171  88
#>  [696,]  696 640 210 170  87
#>  [697,]  697 640 209 169  86
#>  [698,]  698 640 209 169  86
#>  [699,]  699 640 208 168  86
#>  [700,]  700 640 208 168  86
#>  [701,]  701 640 210 167  88
#>  [702,]  702 640 209 166  87
#>  [703,]  703 640 209 166  87
#>  [704,]  704 640 209 166  87
#>  [705,]  705 640 209 166  87
#>  [706,]  706 640 208 165  86
#>  [707,]  707 640 205 164  85
#>  [708,]  708 640 207 164  85
#>  [709,]  709 640 209 163  85
#>  [710,]  710 640 209 163  85
#>  [711,]  711 640 208 162  85
#>  [712,]  712 640 208 162  85
#>  [713,]  713 640 207 161  84
#>  [714,]  714 640 207 161  84
#>  [715,]  715 640 207 161  84
#>  [716,]  716 640 206 160  83
#>  [717,]  717 640 206 160  83
#>  [718,]  718 640 205 159  82
#>  [719,]  719 640 203 159  82
#>  [720,]  720 640 202 158  81
#>  [721,]  721 640 197 161  77
#>  [722,]  722 640 196 163  77
#>  [723,]  723 640 214 153  87
#>  [724,]  724 640 213 152  84
#>  [725,]  725 640 209 152  83
#>  [726,]  726 640 212 153  84
#>  [727,]  727 640 195 158  78
#>  [728,]  728 640 194 157  77
#>  [729,]  729 640 209 160  74
#>  [730,]  730 640 203 154  68
#>  [731,]  731 640 197 150  88
#>  [732,]  732 640 194 147  86
#>  [733,]  733 640 204 152  82
#>  [734,]  734 640 196 145  72
#>  [735,]  735 640 167 150 111
#>  [736,]  736 640 255 251 232
#>  [737,]  737 640 254 254 254
#>  [738,]  738 640 255 255 255
#>  [739,]  739 640 255 255 255
#>  [740,]  740 640 253 253 253
#>  [741,]  741 640 252 252 252
#>  [742,]  742 640 253 253 253
#>  [743,]  743 640 254 254 254
#>  [744,]  744 640 255 255 255
#>  [745,]  745 640 254 254 254
#>  [746,]  746 640 254 254 254
#>  [747,]  747 640 254 254 254
#>  [748,]  748 640 254 254 254
#>  [749,]  749 640 254 254 254
#>  [750,]  750 640 254 254 254
#>  [751,]  751 640 254 254 254
#>  [752,]  752 640 254 254 254
#>  [753,]  753 640 254 254 254
#>  [754,]  754 640 254 254 254
#>  [755,]  755 640 254 254 254
#>  [756,]  756 640 254 254 254
#>  [757,]  757 640 254 254 254
#>  [758,]  758 640 254 254 254
#>  [759,]  759 640 254 254 254
#>  [760,]  760 640 254 254 254
#>  [761,]  761 640 254 254 254
#>  [762,]  762 640 254 254 254
#>  [763,]  763 640 254 254 254
#>  [764,]  764 640 254 254 254
#>  [765,]  765 640 254 254 254
#>  [766,]  766 640 254 254 254
#>  [767,]  767 640 254 254 254
#>  [768,]  768 640 254 254 254
#>  [769,]  769 640 254 254 254
#>  [770,]  770 640 254 254 254
#>  [771,]  771 640 254 254 254
#>  [772,]  772 640 254 254 254
#>  [773,]  773 640 254 254 254
#>  [774,]  774 640 254 254 254
#>  [775,]  775 640 254 254 254
#>  [776,]  776 640 254 254 254
#>  [777,]  777 640 254 254 254
#>  [778,]  778 640 254 254 254
#>  [779,]  779 640 254 254 254
#>  [780,]  780 640 254 254 254
#>  [781,]  781 640 254 254 254
#>  [782,]  782 640 254 254 254
#>  [783,]  783 640 254 254 254
#>  [784,]  784 640 254 254 254
#>  [785,]  785 640 254 254 254
#>  [786,]  786 640 254 254 254
#>  [787,]  787 640 254 254 254
#>  [788,]  788 640 254 254 254
#>  [789,]  789 640 254 254 254
#>  [790,]  790 640 254 254 254
#>  [791,]  791 640 254 254 254
#>  [792,]  792 640 254 254 254
#>  [793,]  793 640 254 254 254
#>  [794,]  794 640 254 254 254
#>  [795,]  795 640 254 254 254
#>  [796,]  796 640 254 254 254
#>  [797,]  797 640 254 254 254
#>  [798,]  798 640 254 254 254
#>  [799,]  799 640 254 254 254
#>  [800,]  800 640 254 254 254
#>  [801,]  801 640 254 254 254
#>  [802,]  802 640 254 254 254
#>  [803,]  803 640 254 254 254
#>  [804,]  804 640 254 254 254
#>  [805,]  805 640 254 254 254
#>  [806,]  806 640 254 254 254
#>  [807,]  807 640 254 254 254
#>  [808,]  808 640 254 254 254
#>  [809,]  809 640 254 254 254
#>  [810,]  810 640 254 254 254
#>  [811,]  811 640 254 254 254
#>  [812,]  812 640 254 254 254
#>  [813,]  813 640 254 254 254
#>  [814,]  814 640 254 254 254
#>  [815,]  815 640 254 254 254
#>  [816,]  816 640 254 254 254
#>  [817,]  817 640 254 255 253
#>  [818,]  818 640 254 255 253
#>  [819,]  819 640 254 255 253
#>  [820,]  820 640 254 255 253
#>  [821,]  821 640 254 255 253
#>  [822,]  822 640 254 255 253
#>  [823,]  823 640 254 255 253
#>  [824,]  824 640 254 255 253
#>  [825,]  825 640 254 255 253
#>  [826,]  826 640 254 255 253
#>  [827,]  827 640 255 255 254
#>  [828,]  828 640 255 255 254
#>  [829,]  829 640 255 255 254
#>  [830,]  830 640 255 255 254
#>  [831,]  831 640 255 255 254
#>  [832,]  832 640 255 255 254
#>  [833,]  833 640 254 255 253
#>  [834,]  834 640 255 255 254
#>  [835,]  835 640 255 255 254
#>  [836,]  836 640 255 255 254
#>  [837,]  837 640 255 255 254
#>  [838,]  838 640 254 255 253
#>  [839,]  839 640 254 255 253
#>  [840,]  840 640 254 255 253
#>  [841,]  841 640 253 254 252
#>  [842,]  842 640 253 254 252
#>  [843,]  843 640 253 254 252
#>  [844,]  844 640 253 254 252
#>  [845,]  845 640 253 254 252
#>  [846,]  846 640 253 254 252
#>  [847,]  847 640 253 254 252
#>  [848,]  848 640 253 254 252
#>  [849,]  849 640 254 254 254
#>  [850,]  850 640 254 254 254
#>  [851,]  851 640 254 254 254
#>  [852,]  852 640 254 254 254
#>  [853,]  853 640 254 254 254
#>  [854,]  854 640 254 254 254
#>  [855,]  855 640 254 254 254
#>  [856,]  856 640 254 254 254
#>  [857,]  857 640 254 254 254
#>  [858,]  858 640 254 254 254
#>  [859,]  859 640 254 254 254
#>  [860,]  860 640 254 254 254
#>  [861,]  861 640 254 254 254
#>  [862,]  862 640 254 254 254
#>  [863,]  863 640 254 254 254
#>  [864,]  864 640 254 254 254
#>  [865,]  865 640 254 254 254
#>  [866,]  866 640 254 254 254
#>  [867,]  867 640 254 254 254
#>  [868,]  868 640 254 254 254
#>  [869,]  869 640 254 254 254
#>  [870,]  870 640 254 254 254
#>  [871,]  871 640 254 254 254
#>  [872,]  872 640 254 254 254
#>  [873,]  873 640 254 254 254
#>  [874,]  874 640 254 254 254
#>  [875,]  875 640 254 254 254
#>  [876,]  876 640 254 254 254
#>  [877,]  877 640 254 254 254
#>  [878,]  878 640 254 254 254
#>  [879,]  879 640 254 254 254
#>  [880,]  880 640 254 254 254
#>  [881,]  881 640 254 254 254
#>  [882,]  882 640 254 254 254
#>  [883,]  883 640 254 254 254
#>  [884,]  884 640 254 254 254
#>  [885,]  885 640 254 254 254
#>  [886,]  886 640 254 254 254
#>  [887,]  887 640 254 254 254
#>  [888,]  888 640 254 254 254
#>  [889,]  889 640 254 254 254
#>  [890,]  890 640 254 254 254
#>  [891,]  891 640 254 254 254
#>  [892,]  892 640 254 254 254
#>  [893,]  893 640 254 254 254
#>  [894,]  894 640 254 254 254
#>  [895,]  895 640 254 254 254
#>  [896,]  896 640 254 254 254
#>  [897,]  897 640 254 254 254
#>  [898,]  898 640 254 254 254
#>  [899,]  899 640 254 254 254
#>  [900,]  900 640 254 254 254
#>  [901,]  901 640 254 254 254
#>  [902,]  902 640 254 254 254
#>  [903,]  903 640 254 254 254
#>  [904,]  904 640 254 254 254
#>  [905,]  905 640 254 254 254
#>  [906,]  906 640 254 254 254
#>  [907,]  907 640 254 254 254
#>  [908,]  908 640 254 254 254
#>  [909,]  909 640 254 254 254
#>  [910,]  910 640 254 254 254
#>  [911,]  911 640 254 254 254
#>  [912,]  912 640 254 254 254
#>  [913,]  913 640 254 254 254
#>  [914,]  914 640 254 254 254
#>  [915,]  915 640 254 254 254
#>  [916,]  916 640 254 254 254
#>  [917,]  917 640 254 254 254
#>  [918,]  918 640 254 254 254
#>  [919,]  919 640 254 254 254
#>  [920,]  920 640 254 254 254
#>  [921,]  921 640 254 254 254
#>  [922,]  922 640 254 254 254
#>  [923,]  923 640 254 254 254
#>  [924,]  924 640 254 254 254
#>  [925,]  925 640 254 254 254
#>  [926,]  926 640 254 254 254
#>  [927,]  927 640 254 254 254
#>  [928,]  928 640 254 254 254
#>  [929,]  929 640 254 254 254
#>  [930,]  930 640 254 254 254
#>  [931,]  931 640 254 254 254
#>  [932,]  932 640 254 254 254
#>  [933,]  933 640 254 254 254
#>  [934,]  934 640 254 254 254
#>  [935,]  935 640 254 254 254
#>  [936,]  936 640 254 254 254
#>  [937,]  937 640 254 254 254
#>  [938,]  938 640 254 254 254
#>  [939,]  939 640 254 254 254
#>  [940,]  940 640 254 254 254
#>  [941,]  941 640 254 254 254
#>  [942,]  942 640 254 254 254
#>  [943,]  943 640 254 254 254
#>  [944,]  944 640 254 254 254
#>  [945,]  945 640 254 254 254
#>  [946,]  946 640 254 254 254
#>  [947,]  947 640 254 254 254
#>  [948,]  948 640 254 254 254
#>  [949,]  949 640 254 254 254
#>  [950,]  950 640 254 254 254
#>  [951,]  951 640 254 254 254
#>  [952,]  952 640 254 254 254
#>  [953,]  953 640 254 254 254
#>  [954,]  954 640 254 254 254
#>  [955,]  955 640 254 254 254
#>  [956,]  956 640 254 254 254
#>  [957,]  957 640 254 254 254
#>  [958,]  958 640 254 254 254
#>  [959,]  959 640 254 254 254
#>  [960,]  960 640 254 254 254
#>  [961,]  961 640 247 252 253
#>  [962,]  962 640 248 255 255
#>  [963,]  963 640 244 249 247
#>  [964,]  964 640 248 255 250
#>  [965,]  965 640 255 254 254
#>  [966,]  966 640 252 248 253
#>  [967,]  967 640 244 253 255
#>  [968,]  968 640 242 251 255
#>  [969,]  969 640 244 246 247
#>  [970,]  970 640 255 253 245
#>  [971,]  971 640 255 253 229
#>  [972,]  972 640 179 116  96
#>  [973,]  973 640 108  41  18
#>  [974,]  974 640 140  61  40
#>  [975,]  975 640 151  55  42
#>  [976,]  976 640 149  57  44
#>  [977,]  977 640 145  59  41
#>  [978,]  978 640 146  60  42
#>  [979,]  979 640 145  59  41
#>  [980,]  980 640 146  60  42
#>  [981,]  981 640 147  61  43
#>  [982,]  982 640 147  61  43
#>  [983,]  983 640 143  59  41
#>  [984,]  984 640 142  58  40
#>  [985,]  985 640 142  58  40
#>  [986,]  986 640 142  58  40
#>  [987,]  987 640 142  58  40
#>  [988,]  988 640 142  58  40
#>  [989,]  989 640 141  56  40
#>  [990,]  990 640 140  55  39
#>  [991,]  991 640 140  55  39
#>  [992,]  992 640 141  56  40
#>  [993,]  993 640 139  56  41
#>  [994,]  994 640 139  56  41
#>  [995,]  995 640 138  55  40
#>  [996,]  996 640 138  55  40
#>  [997,]  997 640 136  54  42
#>  [998,]  998 640 136  54  42
#>  [999,]  999 640 136  54  42
#> [1000,] 1000 640 136  54  42
#> [1001,] 1001 640 136  54  42
#> [1002,] 1002 640 136  54  43
#> [1003,] 1003 640 134  54  43
#> [1004,] 1004 640 134  54  43
#> [1005,] 1005 640 133  53  42
#> [1006,] 1006 640 132  52  41
#> [1007,] 1007 640 132  52  41
#> [1008,] 1008 640 131  51  40
#> [1009,] 1009 640 131  50  42
#> [1010,] 1010 640 133  52  44
#> [1011,] 1011 640 132  53  44
#> [1012,] 1012 640 131  52  43
#> [1013,] 1013 640 131  52  43
#> [1014,] 1014 640 130  51  42
#> [1015,] 1015 640 131  52  43
#> [1016,] 1016 640 131  52  43
#> [1017,] 1017 640 128  51  42
#> [1018,] 1018 640 127  50  41
#> [1019,] 1019 640 127  50  41
#> [1020,] 1020 640 127  50  41
#> [1021,] 1021 640 127  50  41
#> [1022,] 1022 640 127  50  41
#> [1023,] 1023 640 124  49  40
#> [1024,] 1024 640 123  48  39
#> [1025,] 1025 640 123  48  40
#> [1026,] 1026 640 123  48  40
#> [1027,] 1027 640 123  48  40
#> [1028,] 1028 640 123  48  40
#> [1029,] 1029 640 123  48  40
#> [1030,] 1030 640 123  48  40
#> [1031,] 1031 640 122  47  39
#> [1032,] 1032 640 122  47  39
#> [1033,] 1033 640 119  47  39
#> [1034,] 1034 640 120  47  39
#> [1035,] 1035 640 121  47  41
#> [1036,] 1036 640 121  47  41
#> [1037,] 1037 640 121  47  41
#> [1038,] 1038 640 121  47  41
#> [1039,] 1039 640 120  46  40
#> [1040,] 1040 640 121  47  41
#> [1041,] 1041 640 122  51  47
#> [1042,] 1042 640  87  15  15
#> [1043,] 1043 640 163 125 120
#> [1044,] 1044 640 255 249 242
#> [1045,] 1045 640 253 255 249
#> [1046,] 1046 640 246 253 248
#> [1047,] 1047 640 245 255 249
#> [1048,] 1048 640 246 251 252
#> [1049,] 1049 640 247 247 247
#> [1050,] 1050 640 252 252 252
#> [1051,] 1051 640 255 255 255
#> [1052,] 1052 640 251 253 253
#> [1053,] 1053 640 249 254 252
#> [1054,] 1054 640 250 255 254
#> [1055,] 1055 640 251 255 255
#> [1056,] 1056 640 251 254 252
#> [1057,] 1057 640 254 254 254
#> [1058,] 1058 640 254 254 254
#> [1059,] 1059 640 254 254 254
#> [1060,] 1060 640 254 254 254
#> [1061,] 1061 640 254 254 254
#> [1062,] 1062 640 254 254 254
#> [1063,] 1063 640 254 254 254
#> [1064,] 1064 640 254 254 254
#> [1065,] 1065 640 254 254 254
#> [1066,] 1066 640 254 254 254
#> [1067,] 1067 640 254 254 254
#> [1068,] 1068 640 254 254 254
#> [1069,] 1069 640 254 254 254
#> [1070,] 1070 640 254 254 254
#> [1071,] 1071 640 254 254 254
#> [1072,] 1072 640 254 254 254
#> [1073,] 1073 640 254 254 254
#> [1074,] 1074 640 254 254 254
#> [1075,] 1075 640 254 254 254
#> [1076,] 1076 640 254 254 254
#> [1077,] 1077 640 254 254 254
#> [1078,] 1078 640 254 254 254
#> [1079,] 1079 640 254 254 254
#> [1080,] 1080 640 254 254 254
#> [1081,] 1081 640 254 254 254
#> [1082,] 1082 640 254 254 254
#> [1083,] 1083 640 254 254 254
#> [1084,] 1084 640 254 254 254
#> [1085,] 1085 640 254 254 254
#> [1086,] 1086 640 254 254 254
#> [1087,] 1087 640 254 254 254
#> [1088,] 1088 640 254 254 254
#> [1089,] 1089 640 250 255 254
#> [1090,] 1090 640 250 255 254
#> [1091,] 1091 640 249 255 253
#> [1092,] 1092 640 247 254 251
#> [1093,] 1093 640 252 254 254
#> [1094,] 1094 640 255 255 254
#> [1095,] 1095 640 254 255 251
#> [1096,] 1096 640 242 230 228
#> [1097,] 1097 640 104  50  55
#> [1098,] 1098 640  92  28  34
#> [1099,] 1099 640 100  39  43
#> [1100,] 1100 640  98  32  37
#> [1101,] 1101 640 102  36  41
#> [1102,] 1102 640  96  30  35
#> [1103,] 1103 640  98  33  35
#> [1104,] 1104 640  96  32  32
#> [1105,] 1105 640  98  32  37
#> [1106,] 1106 640  98  32  37
#> [1107,] 1107 640  98  32  37
#> [1108,] 1108 640  98  32  37
#> [1109,] 1109 640  96  32  37
#> [1110,] 1110 640  96  32  38
#> [1111,] 1111 640  95  31  37
#> [1112,] 1112 640  95  31  37
#> [1113,] 1113 640  94  30  36
#> [1114,] 1114 640  94  30  36
#> [1115,] 1115 640  94  30  36
#> [1116,] 1116 640  94  30  36
#> [1117,] 1117 640  94  30  36
#> [1118,] 1118 640  93  29  35
#> [1119,] 1119 640  93  29  35
#> [1120,] 1120 640  92  28  34
#> [1121,] 1121 640  93  29  35
#> [1122,] 1122 640  91  27  33
#> [1123,] 1123 640  93  29  35
#> [1124,] 1124 640  95  31  37
#> [1125,] 1125 640  91  29  35
#> [1126,] 1126 640  91  29  35
#> [1127,] 1127 640  90  30  36
#> [1128,] 1128 640  90  30  36
#> [1129,] 1129 640  89  31  36
#> [1130,] 1130 640  89  31  36
#> [1131,] 1131 640  89  31  36
#> [1132,] 1132 640  88  30  35
#> [1133,] 1133 640  86  30  35
#> [1134,] 1134 640  85  29  34
#> [1135,] 1135 640  85  29  34
#> [1136,] 1136 640  85  29  34
#> [1137,] 1137 640  84  30  35
#> [1138,] 1138 640  84  30  35
#> [1139,] 1139 640  84  30  35
#> [1140,] 1140 640  84  30  35
#> [1141,] 1141 640  82  31  35
#> [1142,] 1142 640  81  30  34
#> [1143,] 1143 640  80  29  33
#> [1144,] 1144 640  80  29  33
#> [1145,] 1145 640  80  29  33
#> [1146,] 1146 640  80  29  33
#> [1147,] 1147 640  80  29  33
#> [1148,] 1148 640  80  29  33
#> [1149,] 1149 640  80  29  33
#> [1150,] 1150 640  80  29  33
#> [1151,] 1151 640  80  29  33
#> [1152,] 1152 640  80  29  33
#> [1153,] 1153 640  79  29  33
#> [1154,] 1154 640  79  29  33
#> [1155,] 1155 640  79  29  33
#> [1156,] 1156 640  79  29  33
#> [1157,] 1157 640  77  30  33
#> [1158,] 1158 640  77  30  33
#> [1159,] 1159 640  75  30  33
#> [1160,] 1160 640  75  30  33
#> [1161,] 1161 640  73  30  33
#> [1162,] 1162 640  72  29  32
#> [1163,] 1163 640  70  30  32
#> [1164,] 1164 640  70  30  32
#> [1165,] 1165 640  71  31  33
#> [1166,] 1166 640  71  31  33
#> [1167,] 1167 640  71  31  33
#> [1168,] 1168 640  70  30  32
#> [1169,] 1169 640  64  30  31
#> [1170,] 1170 640  64  30  31
#> [1171,] 1171 640  66  29  31
#> [1172,] 1172 640  68  29  31
#> [1173,] 1173 640  69  29  31
#> [1174,] 1174 640  69  29  31
#> [1175,] 1175 640  68  28  30
#> [1176,] 1176 640  67  28  30
#> [1177,] 1177 640  67  28  30
#> [1178,] 1178 640  63  29  30
#> [1179,] 1179 640  62  30  31
#> [1180,] 1180 640  61  30  31
#> [1181,] 1181 640  61  29  30
#> [1182,] 1182 640  61  29  30
#> [1183,] 1183 640  61  29  30
#> [1184,] 1184 640  60  28  29
#> [1185,] 1185 640  62  27  31
#> [1186,] 1186 640  66  31  35
#> [1187,] 1187 640  28   7   9
#> [1188,] 1188 640 171 157 158
#> [1189,] 1189 640 255 253 252
#> [1190,] 1190 640 248 251 249
#> [1191,] 1191 640 250 255 254
#> [1192,] 1192 640 249 255 254
#> [1193,] 1193 640 249 255 253
#> [1194,] 1194 640 254 255 253
#> [1195,] 1195 640 255 254 253
#> [1196,] 1196 640 254 255 253
#> [1197,] 1197 640 255 254 254
#> [1198,] 1198 640 254 254 254
#> [1199,] 1199 640 250 255 254
#> [1200,] 1200 640 250 255 254
#> [1201,] 1201 640 254 254 254
#> [1202,] 1202 640 254 254 254
#> [1203,] 1203 640 254 254 254
#> [1204,] 1204 640 254 254 254
#> [1205,] 1205 640 254 254 254
#> [1206,] 1206 640 254 254 254
#> [1207,] 1207 640 254 254 254
#> [1208,] 1208 640 254 254 254
#> [1209,] 1209 640 254 254 254
#> [1210,] 1210 640 254 254 254
#> [1211,] 1211 640 254 254 254
#> [1212,] 1212 640 254 254 254
#> [1213,] 1213 640 254 254 254
#> [1214,] 1214 640 254 254 254
#> [1215,] 1215 640 254 254 254
#> [1216,] 1216 640 254 254 254
#> [1217,] 1217 640 254 254 254
#> [1218,] 1218 640 254 254 254
#> [1219,] 1219 640 254 254 254
#> [1220,] 1220 640 254 254 254
#> [1221,] 1221 640 254 254 254
#> [1222,] 1222 640 254 254 254
#> [1223,] 1223 640 254 254 254
#> [1224,] 1224 640 254 254 254
#> [1225,] 1225 640 254 254 254
#> [1226,] 1226 640 254 254 254
#> [1227,] 1227 640 254 254 254
#> [1228,] 1228 640 254 254 254
#> [1229,] 1229 640 254 254 254
#> [1230,] 1230 640 254 254 254
#> [1231,] 1231 640 255 255 255
#> [1232,] 1232 640 255 255 255
#> [1233,] 1233 640 255 255 255
#> [1234,] 1234 640 255 255 255
#> [1235,] 1235 640 255 255 255
#> [1236,] 1236 640 255 255 255
#> [1237,] 1237 640 255 255 255
#> [1238,] 1238 640 255 255 255
#> [1239,] 1239 640 255 255 255
#> [1240,] 1240 640 255 255 255
#> [1241,] 1241 640 255 255 255
#> [1242,] 1242 640 255 255 255
#> [1243,] 1243 640 255 255 255
#> [1244,] 1244 640 255 255 255
#> [1245,] 1245 640 255 255 255
#> [1246,] 1246 640 255 255 255
#> [1247,] 1247 640 255 255 255
#> [1248,] 1248 640 255 255 255
#> [1249,] 1249 640 255 255 255
#> [1250,] 1250 640 255 255 255
#> [1251,] 1251 640 255 255 255
#> [1252,] 1252 640 255 255 255
#> [1253,] 1253 640 255 255 255
#> [1254,] 1254 640 255 255 255
#> [1255,] 1255 640 255 255 255
#> [1256,] 1256 640 255 255 255
#> [1257,] 1257 640 255 255 255
#> [1258,] 1258 640 255 255 255
#> [1259,] 1259 640 255 255 255
#> [1260,] 1260 640 255 255 255
#> [1261,] 1261 640 255 255 255
#> [1262,] 1262 640 255 255 255
#> [1263,] 1263 640 255 255 255
#> [1264,] 1264 640 255 255 255
#> [1265,] 1265 640 255 255 255
#> [1266,] 1266 640 255 255 255
#> [1267,] 1267 640 255 255 255
#> [1268,] 1268 640 255 255 255
#> [1269,] 1269 640 255 255 255
#> [1270,] 1270 640 255 255 255
#> [1271,] 1271 640 255 255 255
#> [1272,] 1272 640 255 255 255
#> [1273,] 1273 640 255 255 255
#> [1274,] 1274 640 255 255 255
#> [1275,] 1275 640 255 255 255
#> [1276,] 1276 640 255 255 255
#> [1277,] 1277 640 255 255 255
#> [1278,] 1278 640 255 255 255
#> [1279,] 1279 640 255 255 255
#> [1280,] 1280 640 255 255 255