Which Pixels are Inside a Contour

pixelsInContour determines the pixels that are inside a specified contour.

Usage

pixelsInContour(contours, id = NULL)

Arguments

contours

A list of two matrices as produced by findContours.

id

An optional vector indicating the identity of the specific contours for which to run the function.

Value

A matrix with 3 columns:

"id": the contour identity (indicates the set of points belonging to the same contour).

"x": the x coordinates of the points inside the contour.

"y": the y coordinates of the points inside the contour.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

dots <- image(system.file("sample_img/dots.jpg", package = "Rvision"))
dots_gray <- changeColorSpace(dots, "GRAY")
dots_bin <- dots_gray < 200
contours <- findContours(dots_bin)
pixelsInContour(contours, id = c(3, 5))
#>          id    x   y
#>     [1,]  3  700 150
#>     [2,]  3  701 150
#>     [3,]  3  702 150
#>     [4,]  3  703 150
#>     [5,]  3  704 150
#>     [6,]  3  705 150
#>     [7,]  3  693 149
#>     [8,]  3  694 149
#>     [9,]  3  695 149
#>    [10,]  3  696 149
#>    [11,]  3  697 149
#>    [12,]  3  698 149
#>    [13,]  3  699 149
#>    [14,]  3  700 149
#>    [15,]  3  701 149
#>    [16,]  3  702 149
#>    [17,]  3  703 149
#>    [18,]  3  704 149
#>    [19,]  3  705 149
#>    [20,]  3  706 149
#>    [21,]  3  707 149
#>    [22,]  3  708 149
#>    [23,]  3  709 149
#>    [24,]  3  710 149
#>    [25,]  3  711 149
#>    [26,]  3  712 149
#>    [27,]  3  713 149
#>    [28,]  3  690 148
#>    [29,]  3  691 148
#>    [30,]  3  692 148
#>    [31,]  3  693 148
#>    [32,]  3  694 148
#>    [33,]  3  695 148
#>    [34,]  3  696 148
#>    [35,]  3  697 148
#>    [36,]  3  698 148
#>    [37,]  3  699 148
#>    [38,]  3  700 148
#>    [39,]  3  701 148
#>    [40,]  3  702 148
#>    [41,]  3  703 148
#>    [42,]  3  704 148
#>    [43,]  3  705 148
#>    [44,]  3  706 148
#>    [45,]  3  707 148
#>    [46,]  3  708 148
#>    [47,]  3  709 148
#>    [48,]  3  710 148
#>    [49,]  3  711 148
#>    [50,]  3  712 148
#>    [51,]  3  713 148
#>    [52,]  3  714 148
#>    [53,]  3  715 148
#>    [54,]  3  716 148
#>    [55,]  3  687 147
#>    [56,]  3  688 147
#>    [57,]  3  689 147
#>    [58,]  3  690 147
#>    [59,]  3  691 147
#>    [60,]  3  692 147
#>    [61,]  3  693 147
#>    [62,]  3  694 147
#>    [63,]  3  695 147
#>    [64,]  3  696 147
#>    [65,]  3  697 147
#>    [66,]  3  698 147
#>    [67,]  3  699 147
#>    [68,]  3  700 147
#>    [69,]  3  701 147
#>    [70,]  3  702 147
#>    [71,]  3  703 147
#>    [72,]  3  704 147
#>    [73,]  3  705 147
#>    [74,]  3  706 147
#>    [75,]  3  707 147
#>    [76,]  3  708 147
#>    [77,]  3  709 147
#>    [78,]  3  710 147
#>    [79,]  3  711 147
#>    [80,]  3  712 147
#>    [81,]  3  713 147
#>    [82,]  3  714 147
#>    [83,]  3  715 147
#>    [84,]  3  716 147
#>    [85,]  3  717 147
#>    [86,]  3  718 147
#>    [87,]  3  719 147
#>    [88,]  3  684 146
#>    [89,]  3  685 146
#>    [90,]  3  686 146
#>    [91,]  3  687 146
#>    [92,]  3  688 146
#>    [93,]  3  689 146
#>    [94,]  3  690 146
#>    [95,]  3  691 146
#>    [96,]  3  692 146
#>    [97,]  3  693 146
#>    [98,]  3  694 146
#>    [99,]  3  695 146
#>   [100,]  3  696 146
#>   [101,]  3  697 146
#>   [102,]  3  698 146
#>   [103,]  3  699 146
#>   [104,]  3  700 146
#>   [105,]  3  701 146
#>   [106,]  3  702 146
#>   [107,]  3  703 146
#>   [108,]  3  704 146
#>   [109,]  3  705 146
#>   [110,]  3  706 146
#>   [111,]  3  707 146
#>   [112,]  3  708 146
#>   [113,]  3  709 146
#>   [114,]  3  710 146
#>   [115,]  3  711 146
#>   [116,]  3  712 146
#>   [117,]  3  713 146
#>   [118,]  3  714 146
#>   [119,]  3  715 146
#>   [120,]  3  716 146
#>   [121,]  3  717 146
#>   [122,]  3  718 146
#>   [123,]  3  719 146
#>   [124,]  3  720 146
#>   [125,]  3  721 146
#>   [126,]  3  682 145
#>   [127,]  3  683 145
#>   [128,]  3  684 145
#>   [129,]  3  685 145
#>   [130,]  3  686 145
#>   [131,]  3  687 145
#>   [132,]  3  688 145
#>   [133,]  3  689 145
#>   [134,]  3  690 145
#>   [135,]  3  691 145
#>   [136,]  3  692 145
#>   [137,]  3  693 145
#>   [138,]  3  694 145
#>   [139,]  3  695 145
#>   [140,]  3  696 145
#>   [141,]  3  697 145
#>   [142,]  3  698 145
#>   [143,]  3  699 145
#>   [144,]  3  700 145
#>   [145,]  3  701 145
#>   [146,]  3  702 145
#>   [147,]  3  703 145
#>   [148,]  3  704 145
#>   [149,]  3  705 145
#>   [150,]  3  706 145
#>   [151,]  3  707 145
#>   [152,]  3  708 145
#>   [153,]  3  709 145
#>   [154,]  3  710 145
#>   [155,]  3  711 145
#>   [156,]  3  712 145
#>   [157,]  3  713 145
#>   [158,]  3  714 145
#>   [159,]  3  715 145
#>   [160,]  3  716 145
#>   [161,]  3  717 145
#>   [162,]  3  718 145
#>   [163,]  3  719 145
#>   [164,]  3  720 145
#>   [165,]  3  721 145
#>   [166,]  3  722 145
#>   [167,]  3  723 145
#>   [168,]  3  724 145
#>   [169,]  3  681 144
#>   [170,]  3  682 144
#>   [171,]  3  683 144
#>   [172,]  3  684 144
#>   [173,]  3  685 144
#>   [174,]  3  686 144
#>   [175,]  3  687 144
#>   [176,]  3  688 144
#>   [177,]  3  689 144
#>   [178,]  3  690 144
#>   [179,]  3  691 144
#>   [180,]  3  692 144
#>   [181,]  3  693 144
#>   [182,]  3  694 144
#>   [183,]  3  695 144
#>   [184,]  3  696 144
#>   [185,]  3  697 144
#>   [186,]  3  698 144
#>   [187,]  3  699 144
#>   [188,]  3  700 144
#>   [189,]  3  701 144
#>   [190,]  3  702 144
#>   [191,]  3  703 144
#>   [192,]  3  704 144
#>   [193,]  3  705 144
#>   [194,]  3  706 144
#>   [195,]  3  707 144
#>   [196,]  3  708 144
#>   [197,]  3  709 144
#>   [198,]  3  710 144
#>   [199,]  3  711 144
#>   [200,]  3  712 144
#>   [201,]  3  713 144
#>   [202,]  3  714 144
#>   [203,]  3  715 144
#>   [204,]  3  716 144
#>   [205,]  3  717 144
#>   [206,]  3  718 144
#>   [207,]  3  719 144
#>   [208,]  3  720 144
#>   [209,]  3  721 144
#>   [210,]  3  722 144
#>   [211,]  3  723 144
#>   [212,]  3  724 144
#>   [213,]  3  725 144
#>   [214,]  3  726 144
#>   [215,]  3  679 143
#>   [216,]  3  680 143
#>   [217,]  3  681 143
#>   [218,]  3  682 143
#>   [219,]  3  683 143
#>   [220,]  3  684 143
#>   [221,]  3  685 143
#>   [222,]  3  686 143
#>   [223,]  3  687 143
#>   [224,]  3  688 143
#>   [225,]  3  689 143
#>   [226,]  3  690 143
#>   [227,]  3  691 143
#>   [228,]  3  692 143
#>   [229,]  3  693 143
#>   [230,]  3  694 143
#>   [231,]  3  695 143
#>   [232,]  3  696 143
#>   [233,]  3  697 143
#>   [234,]  3  698 143
#>   [235,]  3  699 143
#>   [236,]  3  700 143
#>   [237,]  3  701 143
#>   [238,]  3  702 143
#>   [239,]  3  703 143
#>   [240,]  3  704 143
#>   [241,]  3  705 143
#>   [242,]  3  706 143
#>   [243,]  3  707 143
#>   [244,]  3  708 143
#>   [245,]  3  709 143
#>   [246,]  3  710 143
#>   [247,]  3  711 143
#>   [248,]  3  712 143
#>   [249,]  3  713 143
#>   [250,]  3  714 143
#>   [251,]  3  715 143
#>   [252,]  3  716 143
#>   [253,]  3  717 143
#>   [254,]  3  718 143
#>   [255,]  3  719 143
#>   [256,]  3  720 143
#>   [257,]  3  721 143
#>   [258,]  3  722 143
#>   [259,]  3  723 143
#>   [260,]  3  724 143
#>   [261,]  3  725 143
#>   [262,]  3  726 143
#>   [263,]  3  727 143
#>   [264,]  3  677 142
#>   [265,]  3  678 142
#>   [266,]  3  679 142
#>   [267,]  3  680 142
#>   [268,]  3  681 142
#>   [269,]  3  682 142
#>   [270,]  3  683 142
#>   [271,]  3  684 142
#>   [272,]  3  685 142
#>   [273,]  3  686 142
#>   [274,]  3  687 142
#>   [275,]  3  688 142
#>   [276,]  3  689 142
#>   [277,]  3  690 142
#>   [278,]  3  691 142
#>   [279,]  3  692 142
#>   [280,]  3  693 142
#>   [281,]  3  694 142
#>   [282,]  3  695 142
#>   [283,]  3  696 142
#>   [284,]  3  697 142
#>   [285,]  3  698 142
#>   [286,]  3  699 142
#>   [287,]  3  700 142
#>   [288,]  3  701 142
#>   [289,]  3  702 142
#>   [290,]  3  703 142
#>   [291,]  3  704 142
#>   [292,]  3  705 142
#>   [293,]  3  706 142
#>   [294,]  3  707 142
#>   [295,]  3  708 142
#>   [296,]  3  709 142
#>   [297,]  3  710 142
#>   [298,]  3  711 142
#>   [299,]  3  712 142
#>   [300,]  3  713 142
#>   [301,]  3  714 142
#>   [302,]  3  715 142
#>   [303,]  3  716 142
#>   [304,]  3  717 142
#>   [305,]  3  718 142
#>   [306,]  3  719 142
#>   [307,]  3  720 142
#>   [308,]  3  721 142
#>   [309,]  3  722 142
#>   [310,]  3  723 142
#>   [311,]  3  724 142
#>   [312,]  3  725 142
#>   [313,]  3  726 142
#>   [314,]  3  727 142
#>   [315,]  3  728 142
#>   [316,]  3  729 142
#>   [317,]  3  676 141
#>   [318,]  3  677 141
#>   [319,]  3  678 141
#>   [320,]  3  679 141
#>   [321,]  3  680 141
#>   [322,]  3  681 141
#>   [323,]  3  682 141
#>   [324,]  3  683 141
#>   [325,]  3  684 141
#>   [326,]  3  685 141
#>   [327,]  3  686 141
#>   [328,]  3  687 141
#>   [329,]  3  688 141
#>   [330,]  3  689 141
#>   [331,]  3  690 141
#>   [332,]  3  691 141
#>   [333,]  3  692 141
#>  [ reached getOption("max.print") -- omitted 13282 rows ]