Extract or Replace Parts of an Image

Operators acting on Image objects to extract or replace parts.

Usage

# S3 method for Rcpp_Image
[(x, ...)

# S3 method for Rcpp_Image
[(x, i = NULL, j = NULL) <- value

Arguments

x

An Image object.

...

Other arguments passed to [ when extracting parts. In this case, the function treat that image as an R array and will accept/require the same arguments.

i, j

Indices specifying elements to extract or replace. Indices are numeric vectors which values are coerced to integer as by as.integer (and hence truncated towards zero) or logical vectors which are recycled if necessary to match the dimensions of the image.

value

Vector or matrix with the data to replace the pixels. Typically, this is a matrix with the same number of rows as the number of channels in image, similar to that produced by col2rgb.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon[1:100, 1:100, ]
#> , , 1
#> 
#>        [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>   [1,]   97   97   97   97   97   97   97   97   97    98    98    98    98
#>   [2,]   97   97   97   97   97   97   97   97   97    98    98    98    98
#>   [3,]   95   95   95   95   95   95   96   96   96    96    96    96    98
#>   [4,]   95   95   95   95   95   96   98   98   98    98    98    98   100
#>   [5,]   98   98   98   98   98   98   99   99   99    99    99    99   101
#>   [6,]  100  100  100  100  100  100  101  101  101   101   101   101   103
#>   [7,]  100  100  100  100  101  101  102  102  102   102   102   102   104
#>   [8,]  102  102  102  102  102  102  102  103  103   103   105   105   107
#>   [9,]  102  102  103  103  103  103  105  105  106   106   106   107   107
#>  [10,]  103  103  105  105  105  107  107  107  108   108   108   109   109
#>        [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>   [1,]    98    98    98    98    98   100   100   100   101    99    99   100
#>   [2,]    98    98    98    98    98   100   101   101   102    99   100   101
#>   [3,]    98    98    98   100   100   101   101   101   102   102   102   103
#>   [4,]   100   100   101   101   101   102   102   102   104   104   104   105
#>   [5,]   101   102   102   103   103   103   104   104   105   105   105   107
#>   [6,]   103   103   103   104   104   105   107   107   107   108   108   109
#>   [7,]   104   104   105   107   107   108   109   109   109   108   108   109
#>   [8,]   107   108   109   109   109   110   110   110   111   110   110   111
#>   [9,]   107   110   110   111   111   111   112   110   111   111   111   111
#>  [10,]   110   112   111   112   112   113   113   111   113   113   113   112
#>        [,26] [,27] [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35] [,36] [,37]
#>   [1,]   100   101   101   103   103   103   103   102   102   102   103   102
#>   [2,]   101   101   102   104   104   104   104   103   103   103   104   103
#>   [3,]   104   104   104   103   105   103   103   105   105   105   105   103
#>   [4,]   104   105   105   105   106   105   105   107   107   107   108   105
#>   [5,]   107   108   109   107   107   109   109   107   107   107   108   107
#>   [6,]   110   110   110   109   110   112   112   109   110   110   110   110
#>   [7,]   110   110   110   111   113   113   113   113   114   114   114   112
#>   [8,]   111   113   113   114   114   114   115   115   116   116   116   114
#>   [9,]   112   112   113   113   114   114   115   115   115   116   116   116
#>  [10,]   113   113   114   115   115   116   116   116   118   119   119   119
#>        [,38] [,39] [,40] [,41] [,42] [,43] [,44] [,45] [,46] [,47] [,48] [,49]
#>   [1,]   103   103   104   104   105   105   107   105   106   107   108   111
#>   [2,]   104   105   107   107   107   107   108   106   107   108   109   111
#>   [3,]   105   106   106   107   108   108   108   108   109   109   110   110
#>   [4,]   106   107   108   109   110   110   109   110   111   111   111   113
#>   [5,]   108   107   108   109   109   109   109   112   113   113   114   116
#>   [6,]   111   111   111   112   112   112   113   114   115   116   116   120
#>   [7,]   113   114   114   115   115   115   116   116   119   119   119   122
#>   [8,]   115   116   116   118   118   118   119   119   121   121   121   124
#>   [9,]   118   119   119   120   119   121   122   121   122   122   123   125
#>  [10,]   121   121   121   122   122   122   123   122   125   125   125   127
#>        [,50] [,51] [,52] [,53] [,54] [,55] [,56] [,57] [,58] [,59] [,60] [,61]
#>   [1,]   111   112   113   113   115   115   116   117   117   116   117   118
#>   [2,]   111   112   113   115   115   115   116   117   118   116   117   118
#>   [3,]   110   111   113   113   114   114   115   116   116   117   118   118
#>   [4,]   113   113   113   114   115   115   116   116   117   118   120   120
#>   [5,]   116   116   116   116   117   117   118   116   118   119   119   120
#>   [6,]   120   120   120   120   120   120   121   120   120   122   122   123
#>   [7,]   122   122   122   122   122   122   123   122   122   125   125   126
#>   [8,]   124   124   124   124   125   125   127   125   126   127   128   128
#>   [9,]   125   125   125   125   127   127   128   126   126   128   128   130
#>  [10,]   127   127   127   127   128   129   130   128   128   131   131   131
#>        [,62] [,63] [,64] [,65] [,66] [,67] [,68] [,69] [,70] [,71] [,72] [,73]
#>   [1,]   118   120   121   121   122   122   123   123   123   124   124   122
#>   [2,]   120   121   121   122   122   122   123   124   124   125   125   123
#>   [3,]   120   121   121   120   120   121   122   123   123   125   125   125
#>   [4,]   120   121   122   121   122   123   123   125   125   126   126   126
#>   [5,]   121   122   122   122   125   126   127   128   128   129   129   126
#>   [6,]   123   125   126   126   127   128   129   130   130   132   132   130
#>   [7,]   127   128   128   129   130   132   133   134   134   135   135   132
#>   [8,]   129   130   130   133   133   134   134   135   136   136   136   134
#>   [9,]   131   131   131   132   132   133   134   134   134   134   135   135
#>  [10,]   132   132   133   133   133   134   134   134   135   135   135   135
#>        [,74] [,75] [,76] [,77] [,78] [,79] [,80] [,81] [,82] [,83] [,84] [,85]
#>   [1,]   123   125   125   126   127   127   127   128   129   129   130   130
#>   [2,]   123   125   125   126   127   127   128   128   128   129   130   132
#>   [3,]   125   126   126   126   127   127   128   127   127   127   130   131
#>   [4,]   126   126   126   127   128   128   129   127   127   128   131   132
#>   [5,]   126   126   127   127   127   128   130   130   130   131   133   133
#>   [6,]   130   130   130   130   131   131   132   132   132   133   134   134
#>   [7,]   132   132   132   132   132   133   133   133   134   135   135   135
#>   [8,]   134   134   134   134   134   134   134   135   135   135   135   137
#>   [9,]   135   135   135   135   135   135   135   135   137   137   137   137
#>  [10,]   135   135   135   135   137   137   137   137   137   137   138   138
#>        [,86] [,87] [,88] [,89] [,90] [,91] [,92] [,93] [,94] [,95] [,96] [,97]
#>   [1,]   130   132   133   133   133   133   133   133   133   133   133   132
#>   [2,]   132   133   133   133   134   134   134   134   134   134   133   133
#>   [3,]   131   132   132   132   132   132   132   132   132   132   132   132
#>   [4,]   132   132   133   132   133   133   133   133   133   133   133   133
#>   [5,]   133   133   133   133   133   133   133   133   133   134   134   134
#>   [6,]   134   134   134   134   134   134   134   134   134   134   135   135
#>   [7,]   135   137   137   137   137   137   137   137   137   137   137   137
#>   [8,]   137   137   137   137   137   137   137   137   137   137   137   138
#>   [9,]   137   137   137   137   137   137   138   138   138   138   138   139
#>  [10,]   138   138   138   138   138   138   138   138   138   138   139   139
#>        [,98] [,99] [,100]
#>   [1,]   132   132    132
#>   [2,]   133   133    133
#>   [3,]   132   133    133
#>   [4,]   133   133    133
#>   [5,]   134   134    134
#>   [6,]   135   135    135
#>   [7,]   137   137    137
#>   [8,]   138   138    138
#>   [9,]   139   139    139
#>  [10,]   139   139    140
#> 
#>  [ reached getOption("max.print") -- omitted 90 row(s) and 2 matrix slice(s) ]
balloon[1:100, 1:100] <- col2rgb(c("red", "green", "blue", "yellow"))