Thinning

thinning applies a binary blob thinning operation to achieve a skeletonization of a binary Image object.

Usage

thinning(image, method = "Zhang-Suen", target = "new")

Arguments

image

An 8-bit (8U) single-channel, binary Image object.

method

A string indicating the binarization method to use. It can be any of the following:

"Zhang-Suen (the default)"

"Guo-Hall"

target

The location where the results should be stored. It can take 3 values:

"new":

a new Image object is created and the results are stored inside (the default).

"self":

the results are stored back into image (faster but destructive).

An Image object:

the results are stored in another existing Image object. This is fast and will not replace the content of image but will replace that of target. Note that if target does not have the same dimensions, number of channels, and bit depth as image, an error may be thrown.

Value

If target="new", the function returns an Image object. If target="self", the function returns nothing and modifies image in place. If target is an Image object, the function returns nothing and modifies that Image object in place.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

balloon <- image(system.file("sample_img/balloon1.png", package = "Rvision"))
balloon_gray <- changeColorSpace(balloon, "GRAY")
balloon_th <- niBlackThreshold(balloon_gray)
skel <- thinning(balloon_th)