Create an object of class VideoWriter

Function for creating VideoWriter objects.

videoWriter(outputFile, fourcc, fps, height, width, isColor = TRUE,
  api = "ANY")

Arguments

outputFile
An character string corresponding to the path to an output file.

fourcc
A 4-character string corresponding to the fourcc code of the codec to be used. A list of fourcc codes can be obtained at http://www.fourcc.org/codecs.php.

fps
A numeric value corresponding to the framerate of the output video.

height
An integer value corresponding to the height of the video in pixels.

width
An integer value corresponding to the width of the video in pixels.

isColor
A logical indicating whether the output video is a color (default: TRUE) or grayscale (FALSE) video.

api
A character string corresponding to the API to use for reading the video from the file (see Note; default: "ANY").

Value

A VideoWriter object.

Note

Hereafter is a list of all supported APIs. Note that not all APIs will be available on your computer (actually most of them will not be).

"ANY": automatically select an API.

"VFW": Video For Windows.

"V4L", "V4L2": Video For Linux.

"FIREWIRE", "FIREWARE", "IEEE1394", "DC1394", "CMU1394": IEEE 1394 drivers.

"QT": Quicktime.

"UNICAP": Unicap drivers.

"DSHOW": DirectShow.

"PVAPI": PvAPI, Prosilica GigE SDK.

"OPENNI": OpenNI (for Kinect).

"OPENNI_ASUS": OpenNI (for Asus Xtion).

"XIAPI": XIMEA Camera API.

"AVFOUNDATION": AVFoundation framework for iOS and OSX > Lion.

"GIGANETIX": Smartek Giganetix GigEVisionSDK.

"MSMF": Microsoft Media Foundation.

"WINRT": Microsoft Windows Runtime using Media Foundation.

"INTELPERC": Intel Perceptual Computing SDK.

"OPENNI2": OpenNI2 (for Kinect).

"OPENNI2_ASUS": OpenNI2 (for Asus Xtion and Occipital Structure sensors).

"GPHOTO2": gPhoto2 connection.

"GSTREAMER": GStreamer.

"FFMPEG": FFMPEG library.

"IMAGES": OpenCV Image Sequence.

"ARAVIS": Aravis SDK.

"OPENCV_MJPEG": Built-in OpenCV MotionJPEG codec.

"INTEL_MFX": Intel MediaSDK.

Examples

# TODO