Pixel Values Along a Line Segment

improfile finds all pixels intersected by a line segment and returns their coordinates and color values.

Usage

improfile(image, xi, yi, connectivity = 8, left_to_right = FALSE)

Arguments

image

An Image object.

xi, yi

Two-element vectors containing the x and y coordinates of the start and end points of a line segment. The points can lie outside of the image boundaries but then the line will be clipped on the image boundaries.

connectivity

The connectivity neighborhood to decide whether 2 pixels are contiguous. This parameter can take two values:

4: the neighborhood of a pixel are the four pixels located above (north), below (south), to the left (west) and right (east) of the pixel.

8 (the default): the neighborhood of a pixel includes the four 4-neighbors and the four pixels along the diagonal directions (northeast, northwest, southeast, and southwest).

left_to_right

If `TRUE`, the line segment is traversed from the leftmost endpoint to the rightmost endpoint, regardless of the order in which they are specified in xi and yi. Otherwise (the default), the line is traversed in the order in which the points are specified in xi and yi.

Value

A matrix in which the first two columns indicate the coordinates of the points traversed by the line segment, and the other columns indicate the color values at each location.

Author

Simon Garnier, garnier@njit.edu

Examples

dots <- image(system.file("sample_img/dots.jpg", package = "Rvision"))
improfile(dots, c(1, ncol(dots)), c(nrow(dots) / 2, nrow(dots) / 2))
#>            X   Y   B   G   R
#>    [1,]    1 640 255 255 255
#>    [2,]    2 640 255 255 255
#>    [3,]    3 640 255 255 255
#>    [4,]    4 640 255 255 255
#>    [5,]    5 640 255 255 255
#>    [6,]    6 640 255 255 255
#>    [7,]    7 640 255 255 255
#>    [8,]    8 640 255 255 255
#>    [9,]    9 640 255 255 255
#>   [10,]   10 640 255 255 255
#>   [11,]   11 640 255 255 255
#>   [12,]   12 640 255 255 255
#>   [13,]   13 640 255 255 255
#>   [14,]   14 640 255 255 255
#>   [15,]   15 640 255 255 255
#>   [16,]   16 640 255 255 255
#>   [17,]   17 640 255 255 255
#>   [18,]   18 640 255 255 255
#>   [19,]   19 640 255 255 255
#>   [20,]   20 640 255 255 255
#>   [21,]   21 640 255 255 255
#>   [22,]   22 640 255 255 255
#>   [23,]   23 640 255 255 255
#>   [24,]   24 640 255 255 255
#>   [25,]   25 640 255 255 255
#>   [26,]   26 640 255 255 255
#>   [27,]   27 640 255 255 255
#>   [28,]   28 640 255 255 255
#>   [29,]   29 640 255 255 255
#>   [30,]   30 640 255 255 255
#>   [31,]   31 640 255 255 255
#>   [32,]   32 640 255 255 255
#>   [33,]   33 640 255 255 255
#>   [34,]   34 640 255 255 255
#>   [35,]   35 640 255 255 255
#>   [36,]   36 640 255 255 255
#>   [37,]   37 640 255 255 255
#>   [38,]   38 640 255 255 255
#>   [39,]   39 640 255 255 255
#>   [40,]   40 640 255 255 255
#>   [41,]   41 640 255 255 255
#>   [42,]   42 640 255 255 255
#>   [43,]   43 640 255 255 255
#>   [44,]   44 640 255 255 255
#>   [45,]   45 640 255 255 255
#>   [46,]   46 640 255 255 255
#>   [47,]   47 640 255 255 255
#>   [48,]   48 640 255 255 255
#>   [49,]   49 640 255 255 255
#>   [50,]   50 640 255 255 255
#>   [51,]   51 640 255 255 255
#>   [52,]   52 640 255 255 255
#>   [53,]   53 640 255 255 255
#>   [54,]   54 640 255 255 255
#>   [55,]   55 640 255 255 255
#>   [56,]   56 640 255 255 255
#>   [57,]   57 640 255 255 255
#>   [58,]   58 640 255 255 255
#>   [59,]   59 640 255 255 255
#>   [60,]   60 640 255 255 255
#>   [61,]   61 640 255 255 255
#>   [62,]   62 640 255 255 255
#>   [63,]   63 640 255 255 255
#>   [64,]   64 640 255 255 255
#>   [65,]   65 640 255 255 255
#>   [66,]   66 640 255 255 255
#>   [67,]   67 640 255 255 255
#>   [68,]   68 640 255 255 255
#>   [69,]   69 640 255 255 255
#>   [70,]   70 640 255 255 255
#>   [71,]   71 640 255 255 255
#>   [72,]   72 640 255 255 255
#>   [73,]   73 640 255 255 255
#>   [74,]   74 640 255 255 255
#>   [75,]   75 640 255 255 255
#>   [76,]   76 640 255 255 255
#>   [77,]   77 640 255 255 255
#>   [78,]   78 640 255 255 255
#>   [79,]   79 640 254 254 254
#>   [80,]   80 640 255 255 255
#>   [81,]   81 640 255 255 255
#>   [82,]   82 640 255 255 255
#>   [83,]   83 640 255 255 255
#>   [84,]   84 640 255 255 255
#>   [85,]   85 640 255 255 255
#>   [86,]   86 640 255 255 255
#>   [87,]   87 640 255 255 255
#>   [88,]   88 640 255 255 255
#>   [89,]   89 640 255 255 255
#>   [90,]   90 640 255 255 255
#>   [91,]   91 640 255 255 255
#>   [92,]   92 640 255 255 255
#>   [93,]   93 640 255 255 255
#>   [94,]   94 640 255 255 255
#>   [95,]   95 640 255 255 255
#>   [96,]   96 640 255 255 255
#>   [97,]   97 640 255 255 255
#>   [98,]   98 640 255 255 255
#>   [99,]   99 640 255 255 255
#>  [100,]  100 640 255 255 255
#>  [101,]  101 640 255 255 255
#>  [102,]  102 640 255 255 255
#>  [103,]  103 640 255 255 255
#>  [104,]  104 640 255 255 255
#>  [105,]  105 640 255 255 255
#>  [106,]  106 640 255 255 255
#>  [107,]  107 640 255 255 255
#>  [108,]  108 640 255 255 255
#>  [109,]  109 640 255 255 255
#>  [110,]  110 640 255 255 255
#>  [111,]  111 640 255 255 255
#>  [112,]  112 640 255 255 255
#>  [113,]  113 640 255 255 255
#>  [114,]  114 640 255 255 255
#>  [115,]  115 640 255 255 255
#>  [116,]  116 640 255 255 255
#>  [117,]  117 640 255 255 255
#>  [118,]  118 640 255 255 255
#>  [119,]  119 640 255 255 255
#>  [120,]  120 640 255 255 255
#>  [121,]  121 640 255 255 255
#>  [122,]  122 640 255 255 255
#>  [123,]  123 640 255 255 255
#>  [124,]  124 640 255 255 255
#>  [125,]  125 640 255 255 255
#>  [126,]  126 640 255 255 255
#>  [127,]  127 640 255 255 255
#>  [128,]  128 640 255 255 255
#>  [129,]  129 640 255 255 255
#>  [130,]  130 640 255 255 255
#>  [131,]  131 640 255 255 255
#>  [132,]  132 640 255 255 255
#>  [133,]  133 640 255 255 255
#>  [134,]  134 640 255 255 255
#>  [135,]  135 640 255 255 255
#>  [136,]  136 640 255 255 255
#>  [137,]  137 640 255 255 255
#>  [138,]  138 640 255 255 255
#>  [139,]  139 640 255 255 255
#>  [140,]  140 640 255 255 255
#>  [141,]  141 640 255 255 255
#>  [142,]  142 640 255 255 255
#>  [143,]  143 640 255 255 255
#>  [144,]  144 640 255 255 255
#>  [145,]  145 640 250 254 255
#>  [146,]  146 640 251 255 255
#>  [147,]  147 640 253 255 255
#>  [148,]  148 640 253 255 255
#>  [149,]  149 640 253 255 255
#>  [150,]  150 640 253 255 255
#>  [151,]  151 640 255 255 255
#>  [152,]  152 640 255 255 255
#>  [153,]  153 640 255 255 255
#>  [154,]  154 640 255 255 255
#>  [155,]  155 640 255 255 255
#>  [156,]  156 640 255 255 255
#>  [157,]  157 640 255 255 255
#>  [158,]  158 640 255 255 255
#>  [159,]  159 640 255 255 255
#>  [160,]  160 640 255 255 255
#>  [161,]  161 640 247 252 251
#>  [162,]  162 640 250 255 254
#>  [163,]  163 640 244 252 251
#>  [164,]  164 640 246 251 250
#>  [165,]  165 640 242 255 253
#>  [166,]  166 640 238 252 248
#>  [167,]  167 640 240 251 248
#>  [168,]  168 640 242 245 243
#>  [169,]  169 640 253 254 252
#>  [170,]  170 640 255 244 245
#>  [171,]  171 640 126  70  75
#>  [172,]  172 640  72   5  12
#>  [173,]  173 640 100  32  39
#>  [174,]  174 640 105  32  40
#>  [175,]  175 640  90  32  37
#>  [176,]  176 640  88  30  35
#>  [177,]  177 640  99  33  38
#>  [178,]  178 640  99  33  38
#>  [179,]  179 640  98  32  37
#>  [180,]  180 640  98  32  37
#>  [181,]  181 640  99  32  37
#>  [182,]  182 640 100  33  38
#>  [183,]  183 640 100  33  38
#>  [184,]  184 640 101  34  39
#>  [185,]  185 640 101  34  39
#>  [186,]  186 640 102  35  40
#>  [187,]  187 640 102  35  40
#>  [188,]  188 640 102  35  40
#>  [189,]  189 640 104  34  40
#>  [190,]  190 640 104  34  40
#>  [191,]  191 640 104  34  40
#>  [192,]  192 640 104  34  40
#>  [193,]  193 640 103  36  41
#>  [194,]  194 640 102  35  40
#>  [195,]  195 640 102  35  40
#>  [196,]  196 640 102  35  40
#>  [197,]  197 640 102  35  38
#>  [198,]  198 640 102  35  38
#>  [199,]  199 640 105  36  39
#>  [200,]  200 640 105  36  39
#>  [ reached getOption("max.print") -- omitted 1080 rows ]