Set/Get Video Properties
Source: R/StreamClass.R, R/VideoClass.R, R/VideoWriterClass.R, and 2 more

Set or get the values of various properties of the Video or Stream object.

# S3 method for Rcpp_Stream
setProp(obj, property, value)

# S3 method for Rcpp_Stream
getProp(obj, property)

# S3 method for Rcpp_Video
setProp(obj, property, value)

# S3 method for Rcpp_Video
getProp(obj, property)

# S3 method for Rcpp_VideoWriter
setProp(obj, property, value)

# S3 method for Rcpp_VideoWriter
getProp(obj, property)

setProp(obj, property, value)

getProp(obj, property)

Arguments

obj
A Video or Stream object.

property
A character string specifying the name of the property to modify (see details below for a complete list).

value
The new value of the property.

Value

setProp returns TRUE is the property was set successfully. getProp returns a numeric value or a character string depending on property.

Note

Video properties are:

POS_MSEC: Current position of the video file in milliseconds.

POS_FRAMES: 0-based index of the frame to be decoded/captured next.

POS_AVI_RATIO Relative position of the video file: 0=start of the film, 1=end of the film.

FRAME_WIDTH: Width in pixels of the frames in the video stream.

FRAME_HEIGHT: Height in pixels of the frames in the video stream.

FPS: Frame rate in frames per second.

FOURCC: 4-character FOURCC code of the codec

FRAME_COUNT: Number of frames in the video file.

Setting stream properties depends on a lot of things, mainly your operating system, the camera drivers installed on your coputer and the camera itself. As a consequence, setting stream values might not work at all with your installation.

Stream properties are:

FRAME_WIDTH: Width in pixels of the frames in the video stream.

FRAME_HEIGHT: Height in pixels of the frames in the video stream.

BRIGHTNESS: Brightness of the image

CONTRAST: Contrast of the image

SATURATION: Saturation of the image

HUE: Hue of the image

GAIN: Gain of the image

EXPOSURE: Exposure

See also

Video, video, Stream, stream

Examples

# TODO