Set/Get Video Properties

Set or get the values of various properties of the Video or Stream object.

# S3 method for Rcpp_Stream
setProp(x, property, value)

# S3 method for Rcpp_Stream
getProp(x, property)

# S3 method for Rcpp_Video
setProp(x, property, value)

# S3 method for Rcpp_Video
getProp(x, property)

# S3 method for Rcpp_VideoWriter
setProp(x, property, value)

# S3 method for Rcpp_VideoWriter
getProp(x, property)

setProp(x, property, value)

getProp(x, property)

Arguments

x

A Video or Stream object.

property

A character string specifying the name of the property to modify (see details below for a complete list).

value

The new value of the property.

Value

setProp returns TRUE is the property was set successfully. getProp returns a numeric value or a character string depending on property.

Note

Video properties are:

POS_MSEC: Current position of the video file in milliseconds.

POS_FRAMES: 0-based index of the frame to be decoded/captured next.

POS_AVI_RATIO Relative position of the video file: 0=start of the film, 1=end of the film.

FRAME_WIDTH: Width in pixels of the frames in the video stream.

FRAME_HEIGHT: Height in pixels of the frames in the video stream.

FPS: Frame rate in frames per second.

FOURCC: 4-character FOURCC code of the codec

FRAME_COUNT: Number of frames in the video file.

Setting stream properties depends on a lot of things, mainly your operating system, the camera drivers installed on your coputer and the camera itself. As a consequence, setting stream values might not work at all with your installation.

Stream properties are:

FRAME_WIDTH: Width in pixels of the frames in the video stream.

FRAME_HEIGHT: Height in pixels of the frames in the video stream.

BRIGHTNESS: Brightness of the image

CONTRAST: Contrast of the image

SATURATION: Saturation of the image

HUE: Hue of the image

GAIN: Gain of the image

EXPOSURE: Exposure

See also

Author

Simon Garnier, garnier@njit.edu

Examples

balloon <- video(system.file("sample_vid/Balloon.mp4", package = "Rvision"))
getProp(balloon, "FPS")
#> [1] 30
setProp(balloon, "POS_FRAMES", 25)
#> [1] TRUE